% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderConsole}
\alias{AppenderConsole}
\title{AppenderConsole}
\description{
A simple Appender that outputs to the console. If you
have the package \strong{crayon} installed log levels will be coloured by default
(but you can modify this behaviour by passing a custom \link{Layout})
}
\section{Usage}{
\preformatted{x <- AppenderConsole$new(threshold = NA_integer_, layout = LayoutFormat$new(fmt
  = "\%L [\%t] \%m", timestamp_fmt = "\%H:\%M:\%OS3", colors =
  getOption("yog.colors", list())))

x$append(event)
x$clone(deep = FALSE)

x$destination
}
}

\section{Creating a new Appender}{


\describe{
\item{threshold}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{layout}{A \link{Layout}. See examples.}
}
}

\section{Fields}{


\describe{
\item{\code{threshold}}{An \code{integer}. Threshold of the current Appender
(i.e the maximum log level) that this Appender processes)}

\item{\code{layout}}{The \link{Layout} this Appender usses for formatting \link{LogEvent}s}

\item{\code{logger}}{The \link{Logger} this Appender is attached to}

\item{\code{destination}}{A \code{character} scalar giving the output destination
of the Appender as human readable print output.}
}
}

\examples{
# create a new logger with propagate = FALSE to prevent routing to the root
# logger. Please look at the section "Logger Hirarchies" in the package
# vignette for more info.
logger  <- Logger$new("testlogger", propagate = FALSE)

logger$add_appender(AppenderConsole$new())
logger$add_appender(AppenderConsole$new(
  layout = LayoutFormat$new("[\%t] \%c(): [\%n] \%m from user \%u", colors = getOption("yog.colors"))))

# Will output the message twice because we attached two console appenders
logger$warn("A test message")

}
\seealso{
\link{LayoutFormat}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}}, \code{\link{Appender}}
}
\concept{Appenders}
