% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderDt}
\alias{AppenderDt}
\alias{lgr_data}
\title{Log to an In-Memory Data.Table}
\description{
An Appender that outputs to an in-memory \code{data.table}. This requires that
you have the suggested package \strong{data.table} installed. This kind of
Appender is useful for interactive use, and has very little overhead.
}
\section{Custom Fields}{


\code{AppenderDt} supports \link[=LogEvent]{custom fields}, but they have to be
pre-allocated in the \code{prototype} argument. Custom fields that are not
part of the prototype are discarded.

With the default settings, the custom field \code{value} is included in the
\code{data.table} as a list column to store arbitrary \R objects (see example).
It is recommended to use this feature only \code{TRACE} level.
}

\section{Usage}{
\preformatted{
x <- x$new(threshold = NA_integer_, layout = LayoutFormat$new(fmt = "\%L [\%t]
  \%m", timestamp_fmt = "\%H:\%M:\%S", colors = getOption("lgr.colors")),
  prototype = data.table::data.table(.id = NA_integer_, level = NA_integer_,
  timestamp = Sys.time(), caller = NA_character_, msg = NA_character_, value =
  list(list())), buffer_size = 1e+05)

x$append(event)
x$show(n = 20, threshold = NA_integer_)
x$set_threshold(level)
x$set_layout(layout)
x$filter(event)
x$set_filters(filters)

x$data
x$destination
x$layout
x$threshold
x$filters

}
}

\section{Creating a New Appender}{


\describe{
\item{buffer_size}{\code{integer} scalar. Number of rows of the in-memory
\code{data.table}}
\item{prototype}{A prototype \code{data.table}. This is only necessary to set
manually if you use custom \link{LogEvents}. The default prototype already
contains a list column called \code{value} that can hold arbitrary R values
(see examples)}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via the defined \link{Layout}}
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}
}

\examples{
lg <- Logger$new(
  "test",
  appenders = list(memory = AppenderDt$new()),
  threshold = NA,
  parent = NULL  # to prevent routing to root logger for this example
)
lg$debug("test")
lg$error("test")

# Displaying the log
lg$appenders$memory$data
lg$appenders$memory$show()
show_log(target = lg$appenders$memory)

# If you pass a Logger to show_log, it looks for the first MemoryAppender it
# can find.
show_log(target = lg)

# Data tables can store arbitrary R values in list columns. The default
# AppenderDt comes with a single list column called "value" that you can use
lg$info("the iris data frame", value = iris)
lg$appenders$memory$data
head(lg$appenders$memory$data$value[[3]])

}
\seealso{
\link{LayoutFormat}, \link{simple_logging}, \link[data.table:data.table]{data.table::data.table}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderFile}}, \code{\link{AppenderJson}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
