% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderPushbullet}
\alias{AppenderPushbullet}
\title{Log to Pushbullet}
\description{
Send push notifications to pushbullet. This Appender keeps an in-memory
buffer like \link{AppenderBuffer}, so that it can compose a log message based on
the last \code{buffer_size} events instead of just the last one.
}
\section{Usage}{
\preformatted{
x <- x$new(threshold = NA_integer_, push_level = "error",
  layout = LayoutFormat$new(fmt = "\%K \%t> \%m \%f", timestamp_fmt = "\%H:\%M:\%S",
  colors = NULL), buffer_size = 6, apikey = NULL)

x$flush()
x$set_push_level(level)
x$set_apikey(x)
x$append(event)
x$set_buffer_size(x)
x$set_should_flush(x)
x$show(threshold = NA_integer_, n = 20L)
x$set_threshold(level)
x$set_layout(layout)
x$filter(event)
x$set_filters(filters)

x$apikey
x$push_level
x$buffer_size
x$buffered_events
x$data
x$dt
x$destination
x$layout
x$threshold
x$filters

}
}

\section{Fields}{


\describe{
\item{\code{push_level}, \code{set_push_level()}}{\code{integer} or \code{character}
\link[=log_level]{log level}. Push if a LogEvent of that level is registered
}
\item{\code{buffer_size, set_buffer_size(x)}}{\code{integer} scalar. Number of
\link{LogEvents} to buffer}
\item{\code{should_flush(event)}, \code{set_should_flush(x)}}{ A \code{function} with a
single argument \code{event} (a \link{LogEvent}) that must only return either \code{TRUE}
or \code{FALSE}. If the function returns \code{TRUE}, a log message is generated
from the last \code{buffer_size} events and sent to pushbullet}
}



\describe{
\item{\code{buffer_size, set_buffer_size(x)}}{\code{integer} scalar. Number of \link{LogEvents} to buffer}
\item{\code{flush_on_exit, set_flush_on_exit(x)}}{\code{TRUE} or \code{FALSE}: Whether the
buffer should be flushed when the Appender is garbage collected (f.e when
you close \R)}
\item{\code{flush_on_rotate, set_flush_on_rotate}}{\code{TRUE} or \code{FALSE}: Whether
the buffer should be flushed when the Buffer is full (f.e when you close
\R). Setting this to off can have slighly negative perfomance impacts.}
\item{\code{should_flush(event)}, \code{set_should_flush(x)}}{ A \code{function} with a
single argument \code{event} (a \link{LogEvent}) that must only return either \code{TRUE}
or \code{FALSE}. If the function returns \code{TRUE}, flushing of the buffer is
triggered. Defaults to flushing if a \code{FATAL} event is registered }
}



\describe{
\item{\code{dt}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}



\describe{
\item{\code{data}}{Get the log recorded by this \code{Appender} as a
\code{data.frame}}
}



\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} of predicates (functions
that return either \code{TRUE} or \code{FALSE}). If all of these functions evaluate
to \code{TRUE} the LogEvent is passed on. Since
LogEvents have reference semantics, filters can also be abused to modify
LogEvents before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{




\describe{
\item{\code{flush()}}{Manually trigger flushing}
}


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via this Appenders \link{Layout}}
}


\describe{
\item{\code{append(event)}}{Write the \link{LogEvent} \code{event} to a destination}
}


\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
}
}

\seealso{
\link{LayoutFormat}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderDt}}, \code{\link{AppenderFile}},
  \code{\link{AppenderJson}}, \code{\link{AppenderMemory}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
