% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\docType{data}
\name{AppenderRjdbc}
\alias{AppenderRjdbc}
\title{AppenderRjdbc}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppenderRjdbc
}
\description{
Log to a database table with the RJDBC package.
}
\section{Usage}{
\preformatted{x <- AppenderRjdbc$new(conn, table, field_types = c(level = "smallint",
  timestamp = "timestamp", caller = "varchar(1024)", msg = "varchar(1024)"),
  threshold = NA_integer_, layout = LayoutFormat$new(fmt = "\%L [\%t] \%m",
  timestamp_fmt = "\%H:\%M:\%OS3", colors = getOption("yog.colors", list())))

x$show(n = 20, threshold = NA_integer_)
x$append(event)
x$clone(deep = FALSE)

x$data
x$destination
}
}

\section{Creating a new Appender}{


\describe{
\item{conn}{an RJDBC connection}
\item{field_types}{a named \code{character} vector of field names and types.}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via the defined \link{Layout}}
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}
}

\seealso{
\link{LayoutFormat}, \link{simple_logging}, \link[data.table:data.table]{data.table::data.table}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}}, \code{\link{Appender}}
}
\concept{Appenders}
\keyword{datasets}
