% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderSendmail}
\alias{AppenderSendmail}
\title{Send emails via sendmailR}
\description{
Send mails via \code{\link[sendmailR:sendmail]{sendmailR::sendmail()}}, which requires that you have access
to an SMTP server that does not require authentication. This
Appender keeps an in-memory buffer like \link{AppenderBuffer}. If the buffer is
flushed, usually because an event of specified magnitude is encountered, all
buffered events are concatenated to a single message. The default behaviour
is to push the last 30 log events in case a \code{fatal} event is encountered.
}
\note{
The default Layout's \code{fmt} indents each log entry with 3 blanks. This
is a workaround so that Microsoft Outlook does not mess up the line breaks.
}
\section{Usage}{
\preformatted{
x <- AppenderSendmail$new(to, control, threshold = NA_integer_, flush_threshold
  = "fatal", layout = LayoutFormat$new(fmt = " \%L [\%t] \%m \%f", timestamp_fmt =
  "\%H:\%M:\%S"), subject_layout = LayoutFormat$new(fmt = "[LGR] \%L: \%m"),
  buffer_size = 29, from = get_user(), cc = NULL, bcc = NULL, html = FALSE,
  headers = NULL, filters = NULL)

x$add_filter(filter, name = NULL)
x$append(event)
x$filter(event)
x$flush()
x$format(color = FALSE, ...)
x$remove_filter(pos)
x$set_bcc(x)
x$set_buffer_size(x)
x$set_cc(x)
x$set_control(x)
x$set_filters(filters)
x$set_flush_threshold(level)
x$set_from(x)
x$set_headers(x)
x$set_html(x)
x$set_layout(layout)
x$set_should_flush(x)
x$set_subject_layout(layout)
x$set_threshold(level)
x$set_to(x)
x$show(threshold = NA_integer_, n = 20L)

x$bcc
x$buffer_df
x$buffer_dt
x$buffer_events
x$buffer_size
x$cc
x$control
x$data
x$destination
x$dt
x$filters
x$flush_threshold
x$from
x$headers
x$html
x$layout
x$should_flush
x$subject_layout
x$threshold
x$to

}
}

\section{Fields}{


\describe{
\item{\code{headers}, \code{control}}{see \code{\link[sendmailR:sendmail]{sendmailR::sendmail()}}}
}


\describe{
\item{\code{to}, \code{from}, \code{cc}, \code{bcc}}{
\code{character} vectors.
}
\item{\code{html}, \code{set_html()}}{\code{TRUE} or \code{FALSE}. Send a html email message?
This does currently only formats the log contents as monospace verbatim
text.
}
}



\describe{
\item{\code{subject_layout}, \code{set_layout(subject_layout)}}{Like \code{layout}, but
used to format the subject/title of the digest. While \code{layout} is applied
to each LogEvent of the digest, \code{subject_layout} is only applied to
the last one.
}
}



\describe{
\item{\code{buffer_size, set_buffer_size(x)}}{\code{integer} scalar \code{>= 0} Number of
\link{LogEvents} to buffer.
}

\item{\code{buffer_events}, \code{buffer_df}, \code{buffer_dt}}{
The contents of the buffer as a \code{list} of \link[=LogEvent]{LogEvents}, a
\code{data.frame} or a \code{data.table}.
}

\item{\code{flush_threshold}, \code{set_flush_threshold()}}{\code{integer} or \code{character}
\link[=log_level]{log level}. Minimum event level that will trigger flushing of
the buffer. This behaviour is implemented through \code{should_flush()},
and you can modify that function for different behaviour.
}
\item{\code{should_flush(event)}, \code{set_should_flush(x)}}{
A function with exactly one arguments: \code{event}.
If the function returns \code{TRUE}, flushing of the buffer
is triggered. Defaults to flushing if an event of level \code{error}
or higher is registered.}
}



\describe{
\item{\code{dt}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}



\describe{
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.frame}}
}



\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} that may contain
\code{functions} or any \R object with a \code{filter()} method. These functions
must have exactly one argument: \code{event} which will get passed the
LogEvent when the Filterable's \code{filter()} method is invoked.
If all of these functions evaluate to \code{TRUE} the LogEvent is passed on.
Since LogEvents have reference semantics, filters can also be abused to
modify them before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{






\describe{
\item{\code{flush()}}{Manually trigger flushing of the buffer}
}


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formatted for console output
via this Appenders \link{Layout}}
}


\describe{
\item{\code{append(event)}}{Tell the Appender to process a \link{LogEvent} \code{event}.
This method is usually not called by the user, but invoked by a
\link{Logger}
}
}


\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
\item{\code{add_filter(filter, name = NULL)}, \code{remove_filter(pos)}}{
Add or remove a filter. When adding a filter an optional \code{name} can
be specified. \code{remove_filter()} can remove by position or name (if one
was specified)
}
}
}

\seealso{
\link{LayoutFormat}, \link{LayoutGlue}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderFileRotating}},
  \code{\link{AppenderFile}}, \code{\link{AppenderGmail}},
  \code{\link{AppenderJson}},
  \code{\link{AppenderPushbullet}},
  \code{\link{AppenderRjdbc}},
  \code{\link{AppenderSyslog}},
  \code{\link{AppenderTable}}, \code{\link{Appender}}
}
\concept{Appenders}
