% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_list.R
\name{as_event_list}
\alias{as_event_list}
\title{Convert a data.frame to a list of LogEvents}
\usage{
as_event_list(x, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\value{

}
\description{
convert a data.frame as returned for example by AppenderDt$dt to a list of
}
