% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogEvent.R
\name{as_LogEvent}
\alias{as_LogEvent}
\alias{as_LogEvent.list}
\alias{as_LogEvent.data.frame}
\title{Coerce objects to LogEvent}
\usage{
as_LogEvent(x, ...)

\method{as_LogEvent}{list}(x, ...)

\method{as_LogEvent}{data.frame}(x, ...)
}
\arguments{
\item{x}{any supported \R object}

\item{...}{currently ignored}
}
\value{
a \link{LogEvent}
}
\description{
Smartly coerce \R objects that look like LogEvents to LogEvents. Mainly
useful for developing Appenders.
}
\details{
\strong{Note}: \code{as_LogEvent.data.frame()} only supports single-row \code{data.frames}
}
\seealso{
Other docs relevant for extending lgr: 
\code{\link{LogEvent}},
\code{\link{event_list}()},
\code{\link{standardize_threshold}()}
}
\concept{docs relevant for extending lgr}
