% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_config.R
\name{as_logger_config}
\alias{as_logger_config}
\alias{as_logger_config.list}
\alias{as_logger_config.character}
\alias{as_logger_config.Logger}
\title{Create logger configs from different sources}
\usage{
as_logger_config(x)

\method{as_logger_config}{list}(x)

\method{as_logger_config}{character}(x)

\method{as_logger_config}{Logger}(x)
}
\arguments{
\item{x}{any \R object. Especially:
\itemize{
\item A \code{character} scalar. This can either be the path to a
\link[=https://yaml.org/]{yaml} file, or valid yaml directly.
}}
}
\value{
a logger_config object
}
\description{
Coerce any supported \R object to a \code{logger_config} object. You usually do
not have to call this function directly, as it is automatically
invoked by the \code{config()} method of Loggers (see examples)
}
\examples{
cfg <- "
Logger:
  name: test/blubb
  threshold: info
  propagate: false
  appenders:
    AppenderFile:
      file: /tmp/blah.txt
"
lg$config(as_logger_config(cfg))
# but you can just do the following directly
lg <- get_logger("test")
lg$config(cfg)

}
