% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CannotInitializeAbstractClassError}
\alias{CannotInitializeAbstractClassError}
\title{Title}
\usage{
CannotInitializeAbstractClassError(class = parent.frame(2)[["classes"]])
}
\arguments{
\item{class}{}
}
\value{

}
\description{
Title
}
\concept{developer tools}
