% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{default_exception_handler}
\alias{default_exception_handler}
\title{Demote an Exception to a Warning}
\usage{
default_exception_handler(e)
}
\arguments{
\item{e}{a \code{character} scalar (usually a \code{try-error})}
}
\value{
The warning as \code{character} vector
}
\description{
Throws a warning instead of stopping the program.
}
