% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_functions.R
\name{default_should_flush}
\alias{default_should_flush}
\title{Default Should Flush Function}
\usage{
default_should_flush(event)
}
\arguments{
\item{event}{a \link{LogEvent}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
This is the default "shuld flush" trigger function for Appenders that
support such a mechanism, such as \link{AppenderBuffer} and \link{AppenderDbi}. It
returnes \code{TRUE} if the event's \code{level} meets or exceedes the Appender's
\code{flush_threshold}.
}
