% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{.obj}
\alias{.obj}
\title{Retrieve the Parent object of a Filter}
\usage{
.obj()
}
\description{
Only for use within \verb{$filter()} methods of \link[=EventFilter]{EventFilters}
}
\examples{
lg <- get_logger("test")
f <- function(event) {
  cat("via event$.logger:", event$.logger$threshold, "\n")  #  works for loggers only
  cat("via .obj():      ",.obj()$threshold, "\n") # works for loggers and appenders
  TRUE
}
lg$add_filter(f)
lg$fatal("test")
lg$config(NULL)
}
