% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filterable.R, R/Filter.R
\name{Filterable}
\alias{Filterable}
\alias{.obj}
\title{Abstract Class for Filterables}
\usage{
.obj()
}
\description{
\strong{Abstract classes} are exported for package developers that want to extend
them, they cannot be instantiated directly.

Superclass for classes that have a \code{filter()} method such as \link{Appenders} and
\link{Loggers}. This class is only exported for package developers that want to
extend it.
}
\section{Fields}{


\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} that may contain
\code{functions} or any \R object with a \code{filter()} method. These functions
must have exactly one argument: \code{event} which will get passed the
LogEvent when the Filterable's \code{filter()} method is invoked.
If all of these functions evaluate to \code{TRUE} the LogEvent is passed on.
Since LogEvents have reference semantics, filters can also be abused to
modify them before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{

\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
\item{\code{add_filter(filter, name = NULL)}, \code{remove_filter(pos)}}{
Add or remove a filter. When adding a filter an optional \code{name} can
be specified. \code{remove_filter()} can remove by position or name (if one
was specified)
}
}
}

\seealso{
Other abstract classes: 
\code{\link{AppenderMemory}},
\code{\link{AppenderTable}}
}
\concept{abstract classes}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-filter}{\code{Filterable$filter()}}
\item \href{#method-add_filter}{\code{Filterable$add_filter()}}
\item \href{#method-remove_filter}{\code{Filterable$remove_filter()}}
\item \href{#method-set_filters}{\code{Filterable$set_filters()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter"></a>}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filterable$filter(event)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_filter"></a>}}
\subsection{Method \code{add_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filterable$add_filter(filter, name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_filter"></a>}}
\subsection{Method \code{remove_filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filterable$remove_filter(pos)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_filters"></a>}}
\subsection{Method \code{set_filters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Filterable$set_filters(filters)}\if{html}{\out{</div>}}
}

}
}
