% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_caller}
\alias{get_caller}
\title{Get the Calling Function}
\usage{
get_caller(where = -1L)
}
\arguments{
\item{where}{\code{integer} scalar (usually negative). Look up that many frames
up the call stack}
}
\value{
a \code{character} scalar
}
\description{
Tries to determine the calling functions based on \code{where}.
}
\examples{
foo <- function() get_caller(-1L)
foo()


}
\seealso{
\code{\link[base:sys.call]{base::sys.call()}}
}
