% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_user}
\alias{get_user}
\title{Get the Current System User}
\usage{
get_user(fallback = "unknown user")
}
\arguments{
\item{fallback}{A fallback in case the user name could not be determined}
}
\value{
a \code{character} scalar.
}
\description{
Try to determine the current user. Defaults to \code{getOption("yog.user")}. If
the option is not set, \code{Sys.info()[["user"]]} is used. If the option is not
set and the package \strong{whoami} is available, the user name
is guessed based on whichever of the following is available: \code{email_address},
\code{fullname}, \code{gh_username}, \code{username}.
}
\examples{
  get_user()
}
