% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_index.R
\name{initialize_implicit_loggers}
\alias{initialize_implicit_loggers}
\title{Initilize all loggers along all logger paths}
\usage{
initialize_implicit_loggers()
}
\description{
Sometimes necessary for things like \code{logger_index()}
}
\examples{
remove_all_loggers()
get_logger("a/test/logger")
print(ls(loggers))
}
