% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{is_filter}
\alias{is_filter}
\title{Check if an R Object is a Filter}
\usage{
is_filter(x)
}
\arguments{
\item{x}{any \R Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Returns \code{TRUE} for any \R object that can be used as a Filter
for \link{Loggers} or, \link{Appenders}:
\itemize{
\item a \code{function} with the single argument \code{event};
\item an \link{EventFilter} \link[R6:R6Class]{R6::R6} object; or
\item any object with a \verb{$filter(event)} method.
}

\strong{Note:} A Filter \strong{must} return a scalar \code{TRUE} or \code{FALSE}, but this
property cannot be checked by \code{\link[=is_filter]{is_filter()}}.
}
\seealso{
\link{EventFilter}, \link{Filterable}
}
