% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{LayoutTable}
\alias{LayoutTable}
\title{LayoutTable}
\description{
\code{LayoutTable} is an internal class that is only exported for developers that
want to extend yog. You are probably looking for either \link{LayoutDbi} or
\link{LayoutJson}.
}
\section{Creating a New Layout}{


\describe{
\item{\code{event_vals}}{Names of the fields of the \link{LogEvent}
to include in the output object.
}

\item{\code{logger_vals}}{Names of the fields of the \link{Logger} that produced the
\link{LogEvent} to include in the output object.
}

\item{\code{other_vals}}{A named \code{list} of any kind of \R values. Functions in
this `list`` will be executed with no arguments and their results will
be included in the results object (see examples)
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\seealso{
Other Layouts: \code{\link{LayoutDbi}},
  \code{\link{LayoutFormat}}, \code{\link{LayoutJson}},
  \code{\link{Layout}}
}
\concept{Layouts}
