% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_config.R
\name{logger_config}
\alias{logger_config}
\title{Create a Logger configuration object}
\usage{
logger_config(appenders = list(), threshold = NULL, filters = list(),
  exception_handler = default_exception_handler, propagate = TRUE)
}
\arguments{
\item{propagate}{}
}
\value{
a \code{list} with subclass \code{"logger_config"}
}
\description{
Create a Logger configuration object
}
\examples{
# call without arguments to generate the default configuration
cfg <- logger_config()

}
