% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_index.R
\name{logger_index}
\alias{logger_index}
\title{Return a data.frame of all registered loggers}
\usage{
logger_index()
}
\value{
a \code{logger_index} \code{data.frame}
}
\description{
Return a data.frame of all registered loggers
}
\examples{
get_logger("tree/leaf")
get_logger("shrub/leaf")
get_logger("plant/shrub/leaf")
logger_index()
}
\seealso{
\code{\link[=logger_tree]{logger_tree()}} for a more visual representation of registered
loggers
}
