% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_tree.R
\name{logger_tree}
\alias{logger_tree}
\title{Logger Tree}
\usage{
logger_tree()
}
\value{
\code{data.frame} with subclass \code{"logger_tree"}
}
\description{
Logger Tree
}
\examples{
get_logger("fancymodel")
get_logger("fancymodel/shiny")$
  set_propagate(FALSE)

get_logger("fancymodel/shiny/ui")$
  set_appenders(AppenderConsole$new())

get_logger("fancymodel/shiny/server")$
  set_appenders(list(AppenderConsole$new(), AppenderConsole$new()))$
  set_threshold("trace")

get_logger("fancymodel/plumber")

if (requireNamespace("cli")){
  logger_tree()
}
}
