% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\docType{data}
\name{LoggerRoot}
\alias{LoggerRoot}
\title{Special logger subclass for the root logger. Currently exactly like a
normal Logger, but prevents the threshold to be set to NULL}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LoggerRoot
}
\description{
Special logger subclass for the root logger. Currently exactly like a
normal Logger, but prevents the threshold to be set to NULL
}
\keyword{datasets}
