% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_levels.R
\name{named_union}
\alias{named_union}
\title{union for named vectors. names of \code{y} override names of \code{x}}
\usage{
named_union(x, y)
}
\description{
union for named vectors. names of \code{y} override names of \code{x}
}
