% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatting.R
\name{pad_right}
\alias{pad_right}
\alias{pad_left}
\title{Pad Character Vectors}
\arguments{
\item{x}{a \code{character} vector}

\item{width}{\code{integer} scalar. target string width}

\item{pad}{\code{character} scalar. the symbol to pad with}
}
\description{
Pad Character Vectors
}
\examples{
pad_left("foo", 5)
pad_right("foo", 5, ".")
pad_left(c("foo", "foooooo"), pad = ".")
}
