% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Logger.R
\name{print.ancestry}
\alias{print.ancestry}
\alias{format.ancestry}
\title{Print an Ancenstry Object}
\usage{
\method{print}{ancestry}(x, color = requireNamespace("crayon", quietly =
  TRUE), ...)

\method{format}{ancestry}(x, color = requireNamespace("crayon", quietly =
  TRUE), ...)
}
\arguments{
\item{x}{any \R Object}

\item{color}{\code{TRUE} or \code{FALSE}: Output with color? Requires the Package
\strong{crayon}}

\item{...}{ignored}
}
\value{
\code{x} (invisibly)
}
\description{
Print an Ancenstry Object
}
