% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Appender.R
\name{print.Appender}
\alias{print.Appender}
\title{Print an Appender object}
\usage{
\method{print}{Appender}(x, color = requireNamespace("crayon", quietly =
  TRUE), ...)
}
\arguments{
\item{x}{any \R Object}

\item{color}{\code{TRUE} or \code{FALSE}: Output with color? Requires the Package
\strong{crayon}}

\item{...}{ignored}
}
\value{
\code{print()} returns \code{x} (invisibly), \code{format()} returns a \code{character} vector.
}
\description{
The \code{print()} method for Loggers displays the most important aspects of
the Appender.
}
\examples{
# print most important details of logger
print(lgr$console)
}
