% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_LogEvent.R
\name{print.LogEvent}
\alias{print.LogEvent}
\alias{format.LogEvent}
\title{Print or Format Logging Data}
\usage{
\method{print}{LogEvent}(x, fmt = "\%L [\%t] \%m  \%f",
  timestamp_fmt = "\%Y-\%m-\%d \%H:\%M:\%S",
  colors = getOption("lgr.colors"),
  log_levels = getOption("lgr.log_levels"), pad_levels = "right",
  user = x$user, ...)

\method{format}{LogEvent}(x, fmt = "\%L [\%t] \%m  \%f",
  timestamp_fmt = "\%Y-\%m-\%d \%H:\%M:\%S", colors = NULL,
  log_levels = getOption("lgr.log_levels"), pad_levels = "right",
  user = x$user, ...)
}
\arguments{
\item{x}{a \link{LogEvent} or \link{lgr_data} Object}

\item{fmt}{A \code{character} scalar that may contain any of the tokens listed
bellow in the section Format Tokens.}

\item{timestamp_fmt}{see \code{\link[=format.POSIXct]{format.POSIXct()}}}

\item{colors}{A \code{list} of \code{functions} that will be used to color the
log levels (likely from \link[crayon:crayon]{crayon::crayon}).}

\item{log_levels}{a named \code{integer} vector of log levels.}

\item{pad_levels}{\code{right}, \code{left} or \code{NULL}. Whether or not to pad the log
level names to the same width on the left or right side, or not at all.}

\item{user}{The user}

\item{...}{ignored}
}
\value{
\code{x} for \code{print()} and a \code{character} scalar for \code{format()}
}
\description{
Print or Format Logging Data
}
\section{Format Tokens}{

\describe{
\item{\code{\%t}}{A timestamp (see also \code{timestamp_fmt}) }
\item{\code{\%l}}{the log level}
\item{\code{\%L}}{the log level (uppercase)}
\item{\code{\%n}}{the log level (numeric)}
\item{\code{\%u}}{the current user}
\item{\code{\%p}}{the PID (process ID). Useful when logging code that uses
multiple threads.}
\item{\code{\%c}}{the calling function}
\item{\code{\%m}}{the log message}
}
}

\examples{
x <- LogEvent$new(level = 300, msg = "a test event", logger = lgr)
print(x)
print(x, colors = NULL)


}
