% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Logger.R
\name{print.Logger}
\alias{print.Logger}
\alias{format.Logger}
\title{Print a Logger Object}
\usage{
\method{print}{Logger}(x, colors = requireNamespace("crayon", quietly =
  TRUE), ...)

\method{format}{Logger}(x, colors = requireNamespace("crayon", quietly =
  TRUE), ...)
}
\arguments{
\item{x}{any \R Object}

\item{colors}{\code{TRUE} or \code{FALSE}: Output with colors? Requires the Package
\strong{crayon}}

\item{...}{ignored}
}
\value{
\code{print()} returns \code{x} (invisibly), \code{format()} returns a \code{character} vector.
}
\description{
Display a short summary of the most important aspects of the logger
}
\examples{
print(yog)
}
