% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_tree.R
\name{print.logger_tree}
\alias{print.logger_tree}
\title{Print Logger Trees}
\usage{
\method{print}{logger_tree}(x, ...)
}
\arguments{
\item{x}{a \link[=logger_tree]{logger_tree}}

\item{...}{pased on to \link[cli:tree]{cli::tree}#'}
}
\value{
\code{x} (invisibly)
}
\description{
Print Logger Trees
}
