% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json_log.R
\name{read_json_lines}
\alias{read_json_lines}
\title{Read a JSON logfile}
\usage{
read_json_lines(file, ...)
}
\arguments{
\item{file}{\code{character} scalar. path to a JSON logfile (one JSON object per line)}

\item{...}{passed on to \code{\link[jsonlite:steam_in]{jsonlite::steam_in()}}}
}
\value{
a \code{data.frame}
}
\description{
Read a JSON logfile
}
\seealso{
\link{LayoutJson}
}
