% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{select_dbi_layout}
\alias{select_dbi_layout}
\title{Select Appropriate Database Table Layout}
\usage{
select_dbi_layout(conn, table)
}
\arguments{
\item{conn}{a \link[DBI:dbConnect]{DBI connection}}

\item{table}{a \code{character} scalar. The name of the table to log to.}
}
\description{
Selects an appropriate Layout for a database table based on
a DBI connection and - if it already exists in the database -
the table itself.
}
