% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_logging.R
\name{simple_logging}
\alias{simple_logging}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\alias{log_exception}
\alias{threshold}
\alias{add_appender}
\alias{remove_appender}
\alias{show_log}
\title{Simple Logging}
\usage{
FATAL(msg, ...)

ERROR(msg, ...)

WARN(msg, ...)

INFO(msg, ...)

DEBUG(msg, ...)

TRACE(msg, ...)

log_exception(...)

threshold(level, target = yog::yog)

add_appender(appender, target = yog::yog)

remove_appender(pos, target = yog::yog)

show_log(n = 20, threshold = NA, target = yog::yog)
}
\arguments{
\item{msg, ...}{passed on to \code{\link[base:sprintf]{base::sprintf()}}}

\item{level, threshold}{an \code{integer} or \code{character} scalar, see
\code{getOption("yog.log_levels")} for possible values. For \code{threshold} \code{0}
(\code{"off"}) and \code{NA} (\code{"all"}) are also valid.}

\item{target}{a \link{Logger} or \link{Appender}. Defaults to the root logger.}

\item{appender}{an \code{Appender}}

\item{pos}{\code{integer} index or \code{character} names of the appenders to remove}

\item{n}{\code{integer} scalar. Show only the last \code{n} log entries that match
\code{threshold}}
}
\value{
\code{FATAL()} ... \code{TRACE()} and \code{log_exception()} return the log message
as a \code{character} vector.

\code{threshold()} returns the log level of \code{target} as \code{integer} (invisibly)

\code{add_appender()} and \code{remove_appender()} return \code{target}
}
\description{
yog provides some functions for managing loggers that default to the root
logger as target. These are useful if you just use the root logger and
do not want to worry about

\code{threshold()} sets or retrieves the threshold for an appender or logger
(the minimum level of log messages it processes).

\code{add_appender()} and \code{remove_appender()} add \link{Appenders} to \link{Loggers} and
other Appenders.
}
\examples{
FATAL("This is an important message about \%s going wrong", "something")
DEBUG("Debug messages are hidden by default")
threshold("debug")  # you must use lower case names here
DEBUG("Unless we lower the threshold")

}
