% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_levels.R
\name{standardize_threshold}
\alias{standardize_threshold}
\title{Standardize User-Input Log Levels to Their Integer Representation}
\usage{
standardize_threshold(
  x,
  log_levels = c(getOption("lgr.log_levels"), c(all = NA_integer_, off = 0L)),
  allow_null = FALSE
)
}
\arguments{
\item{x}{a \code{character} or \code{integer} scalar, or vector for
standardize_log_levels}

\item{log_levels}{A named character vector of valid log levels}
}
\value{
An unnamed \code{integer} vector
}
\description{
Standardize User-Input Log Levels to Their Integer Representation
}
