% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{suspend_logging}
\alias{suspend_logging}
\alias{unsuspend_logging}
\alias{without_logging}
\title{Suspend All Logging}
\usage{
suspend_logging()

unsuspend_logging()

without_logging(code)
}
\arguments{
\item{code}{Any \R code}
}
\value{
\code{suspend_logging()} and \code{unsuspend_logging()} return \code{NULL} (invisibly),
\code{without_logging} returns whatever \code{code} returns
}
\description{
Completely disable logging for all loggers. This is for example useful for
automated test code.
}
\examples{
without_logging({
  FATAL("FOO")
  INFO("BAR")
})

}
