% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{system_infos}
\alias{system_infos}
\alias{get_user}
\alias{get_caller}
\title{Information About the System}
\usage{
get_user(fallback = "unknown user")

get_caller(where = -1L)
}
\arguments{
\item{fallback}{A fallback in case the user name could not be determined}

\item{where}{\code{integer} scalar (usually negative). Look up that many frames
up the call stack}
}
\value{
a \code{character} scalar.
}
\description{
\code{get_user()} tries to determine the current user. Defaults to
\code{getOption("yog.user")}. If the option is not set, \code{Sys.info()[["user"]]}
is used. If the option is not set and the package \strong{whoami} is available,
the user name is guessed based on whichever of the following is available:
\code{email_address}, \code{fullname}, \code{gh_username}, \code{username}.
}
\examples{
get_user()

foo <- function() get_caller(-1L)
foo()
}
\seealso{
\code{\link[whoami:whoami]{whoami::whoami()}}

\code{\link[base:sys.call]{base::sys.call()}}
}
