% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogEvent.R
\name{toString.LogEvent}
\alias{toString.LogEvent}
\title{Convert a LogEvent to a character string}
\usage{
\method{toString}{LogEvent}(x, ...)
}
\arguments{
\item{x}{a \link{LogEvent}}

\item{...}{ignored}
}
\value{
a \code{character} scalar
}
\description{
Convert a LogEvent to a character string
}
\examples{
toString(LogEvent$new(logger = lgr::lgr))
}
