% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logging.R
\name{with_log_level}
\alias{with_log_level}
\title{Override Log Level}
\usage{
with_log_level(level, code, logger = yog::yog)
}
\arguments{
\item{level}{\code{integer} or \code{character} scalar: the desired log level}

\item{code}{Any \R code}

\item{logger}{a \link{Logger}. defaults to the root logger (yog::yog).}
}
\description{
Temporarily override the log level of all \link{LogEvents} created by target
\link{Logger}
}
\details{
This abuses yog's \link{filter} mechanic to modify LogEvents in-place before it
is passed on the appenders.
}
\examples{
with_log_level("warn", {
  yog$info("More important than it seems")
  FATAL("Really not so bad")
})
}
