% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yog-package.R
\docType{package}
\name{yog-package}
\alias{yog}
\alias{yog-package}
\title{A Fully Featured Logging Framework for R}
\description{
A flexible R6 based logging framework. Yog is under
    heavy development right now, please don't heistate to contact the
    maintainer with bug reports, wishes, suggestions, etc...
}
\section{Options}{


You can also set these options in your \code{.Rprofile} to make them permanent

\describe{
\item{\code{yog.colors}}{a \code{list} of \code{functions} used for coloring the log
levels in console output. Usually these will be functions from the
package \strong{crayon}}
\item{\code{yog.log_levels}}{A named \code{integer} vector of log levels that are
known to yog for labeling, setting thresholds, etc... . Instead of
modifying this option manually use \code{\link[=add_log_levels]{add_log_levels()}} and
\code{\link[=remove_log_levels]{remove_log_levels()}}}
\item{\code{yog.suspend_logging}}{\code{TRUE} or \code{FALSE}. Suspend all logging for
all loggers.  Instead of modifying this option manually use
\code{\link[=suspend_logging]{suspend_logging()}} and \code{\link[=unsuspend_logging]{unsuspend_logging()}}}
\item{\code{yog.user}}{a \code{character} scalar. The default username for new
Loggers. Users can also be set per Logger, though this rarely makes
sense.
}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/s-fleck/yog}
  \item Report bugs at \url{https://github.com/s-fleck/dint/yog}
}

}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (0000-0003-3344-9851)

}
\keyword{internal}
