% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{Appender}
\alias{Appender}
\alias{Appenders}
\title{Appenders}
\description{
Appenders are assigned to \link{Loggers} and manage the output of the \link{LogEvents}
to a destination (such as the console or a text file). An Appender has a
single \link{Layout} that tells it how to format the LogEvent. For details
please refer to the documentations of the specific Appenders.

The Appender class itself \strong{is not designed for direct usage}, but it is the
basis on which all other Appenders are built. Please see the \strong{see also}
section towards the end of this document a list of available Appenders.

More appenders can be found in the package \href{https://github.com/s-fleck/lgrExtra}{lgrExtra}
}
\section{Usage}{
\preformatted{
x <- Appender$new(layout = Layout$new(), threshold = NA_integer_)

x$add_filter(filter, name = NULL)
x$append(event)
x$filter(event)
x$format(color = FALSE, ...)
x$remove_filter(pos)
x$set_filters(filters)
x$set_layout(layout)
x$set_threshold(level)

x$destination
x$filters
x$layout
x$threshold

}
}

\section{Creating a New Appender}{


New Appenders are instantiated with \verb{<AppenderSubclass>$new()}. For the
arguments to \code{new()} please refer to the section \emph{Fields}. You can also
modify those fields after the Appender has been created with setters in the
form of \verb{appender$set_<fieldname>(value)}
}

\section{Fields}{


\describe{
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}
}

\section{Methods}{

\describe{
\item{\code{append(event)}}{Tell the Appender to process a \link{LogEvent} \code{event}.
This method is usually not called by the user, but invoked by a
\link{Logger}
}
}
}

\seealso{
Other Appenders: 
\code{\link{AppenderBuffer}},
\code{\link{AppenderConsole}},
\code{\link{AppenderFileRotating}},
\code{\link{AppenderFile}},
\code{\link{AppenderJson}},
\code{\link{AppenderSyslog}},
\code{\link{AppenderTable}}
}
\concept{Appenders}
\keyword{internal}
\section{Super class}{
\code{\link[lgr:Filterable]{lgr::Filterable}} -> \code{Appender}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Appender$new()}}
\item \href{#method-append}{\code{Appender$append()}}
\item \href{#method-set_threshold}{\code{Appender$set_threshold()}}
\item \href{#method-set_layout}{\code{Appender$set_layout()}}
\item \href{#method-format}{\code{Appender$format()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="add_filter">}\href{../../lgr/html/Filterable.html#method-add_filter}{\code{lgr::Filterable$add_filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="filter">}\href{../../lgr/html/Filterable.html#method-filter}{\code{lgr::Filterable$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="remove_filter">}\href{../../lgr/html/Filterable.html#method-remove_filter}{\code{lgr::Filterable$remove_filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="set_filters">}\href{../../lgr/html/Filterable.html#method-set_filters}{\code{lgr::Filterable$set_filters()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Appender$new(layout = Layout$new(), threshold = NA_integer_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Appender$append(event)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_threshold"></a>}}
\subsection{Method \code{set_threshold()}}{
Set the minimum log level that triggers this Appender. See
\code{\link[=threshold]{threshold()}} for examples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Appender$set_threshold(level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{\code{character} or \code{integer} scalar log level. See \link{log_levels}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_layout"></a>}}
\subsection{Method \code{set_layout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Appender$set_layout(layout)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Appender$format(color = FALSE, ...)}\if{html}{\out{</div>}}
}

}
}
