% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{Appender}
\alias{Appender}
\alias{Appenders}
\title{Appenders}
\description{
Appenders are assigned to \link{Loggers} and manage the output of the \link{LogEvents}
to a destination, such as the console or a text file. An appender must have
a single \link{Layout} that tells it how to format the LogEvent. For details
please refer to the documentations of the specific Appenders.
}
\section{Usage}{
\preformatted{Appender$new(layout = Layout$new(), threshold = NA_integer_)
# public methods
    app$append(event)
    app$set_threshold(level)
    app$set_logger(logger)
    app$set_layout(layout)
# public fields
    app$
# active bindings
    app$threshold
    app$layout
    app$logger
    app$destination
}
}

\section{Creating a new Appender}{


\describe{
\item{threshold}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{layout}{A \link{Layout}. See examples.}
}
}

\seealso{
Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}}
}
\concept{Appenders}
