% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{Appender}
\alias{Appender}
\alias{Appenders}
\title{Abstract Class for Appenders}
\description{
Abstract classes are exported for developers that want to extend them, they
are not useful to casual users. Appenders are assigned to \link{Loggers} and
manage the output of the \link{LogEvents} to a destination, such as the console or
a text file. An appender must have a single \link{Layout} that tells it how to
format the LogEvent. For details please refer to the documentations of the
specific Appenders.
}
\section{Usage}{
\preformatted{x <- Appender$new(layout = Layout$new(), threshold = NA_integer_)

x$append(event)
x$set_threshold(level)
x$set_layout(layout)

x$destination
x$layout
x$threshold
}
}

\section{Creating a New Appender}{


\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{append(event)}}{Write the \link{LogEvent} \code{event} to a destination}
\item{\code{threshold}, \code{set_threshold(level)}}{get/set the \code{Appender} threshold}
\item{\code{layout}, \code{set_layout(layout)}}{get/set the \code{Layout}}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filter(event)}}{Check if an event passes all filters}
\item{\code{filters},\code{set_filters(filters)}}{get/set a list of \code{filters}. A
filter is a function that takes exactly two named arguments: the LogEvent
\code{event} and the Filterable \code{obj} (usually a Logger or Appender)}
}
}

\seealso{
Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderDt}}, \code{\link{AppenderFile}},
  \code{\link{AppenderJson}}, \code{\link{AppenderRjdbc}},
  \code{\link{AppenderTable}}
}
\concept{Appenders}
\keyword{internal}
