% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{Appender}
\alias{Appender}
\alias{Appenders}
\title{Appenders}
\description{
Appenders are assigned to \link{Loggers} and manage the output of the \link{LogEvents}
to a destination, such as the console or a text file. An appender must have
a single \link{Layout} that tells it how to format the LogEvent. For details
please refer to the documentations of the specific Appenders.
}
\section{Usage}{
\preformatted{x <- Appender$new(layout = Layout$new(), threshold = NA_integer_)

x$append(event)
x$set_threshold(level)
x$set_logger(logger)
x$set_layout(layout)

x$destination
x$layout
x$logger
x$threshold
}
}

\section{Creating a new Appender}{


\describe{
\item{threshold}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{layout}{A \link{Layout}. See examples.}
}
}

\section{Fields}{

}

\seealso{
Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}}
}
\concept{Appenders}
