% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{Appender}
\alias{Appender}
\alias{Appenders}
\title{Appenders}
\description{
Appenders are assigned to \link{Loggers} and manage the output of the \link{LogEvents}
to a destination (such as the console or a text file). An Appender has a
single \link{Layout} that tells it how to format the LogEvent. For details
please refer to the documentations of the specific Appenders.

The Appender class itself \strong{is not designed for direct usage}, but it is the
basis on which all other Appenders are built. Please see the \strong{see also}
section towards the end of this document a list of available Appenders.

More appenders can be found in the package \href{https://github.com/s-fleck/lgrExtra}{lgrExtra}
}
\section{Usage}{
\preformatted{
x <- Appender$new(layout = Layout$new(), threshold = NA_integer_)

x$add_filter(filter, name = NULL)
x$append(event)
x$filter(event)
x$format(color = FALSE, ...)
x$remove_filter(pos)
x$set_filters(filters)
x$set_layout(layout)
x$set_threshold(level)

x$destination
x$filters
x$layout
x$threshold

}
}

\section{Creating a New Appender}{


New Appenders are instantiated with \verb{<AppenderSubclass>$new()}. For the
arguments to \code{new()} please refer to the section \emph{Fields}. You can also
modify those fields after the Appender has been created with setters in the
form of \verb{appender$set_<fieldname>(value)}
}

\section{Fields}{


\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}
}

\section{Methods}{

\describe{
\item{\code{append(event)}}{Tell the Appender to process a \link{LogEvent} \code{event}.
This method is usually not called by the user, but invoked by a
\link{Logger}
}
}
}

\seealso{
Other Appenders: 
\code{\link{AppenderBuffer}},
\code{\link{AppenderConsole}},
\code{\link{AppenderFileRotating}},
\code{\link{AppenderFile}},
\code{\link{AppenderJson}},
\code{\link{AppenderSyslog}},
\code{\link{AppenderTable}}
}
\concept{Appenders}
\keyword{internal}
