% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{Appender}
\alias{Appender}
\alias{Appenders}
\title{Abstract Class for Appenders}
\description{
\strong{Abstract classes} are exported for package developers that want to extend
them, they cannot be instantiated directly.

Appenders are assigned to \link{Loggers} and
manage the output of the \link{LogEvents} to a destination, such as the console or
a text file. An Appender must have a single \link{Layout} that tells it how to
format the LogEvent. For details please refer to the documentations of the
specific Appenders.
}
\section{Usage}{
\preformatted{
x <- x$new(layout = Layout$new(), threshold = NA_integer_)

x$append(event)
x$set_threshold(level)
x$set_layout(layout)
x$filter(event)
x$set_filters(filters)

x$destination
x$layout
x$threshold
x$filters

}
}

\section{Creating a New Appender}{


New Appenders are instantiated with \code{<AppenderSubclass>$new()}. For the
arguments to \code{new()} please refer to the section \emph{Fields}. You can also
modify those fields after the Appender has been created with setters in the
form of \code{appender$set_<fieldname>(value)}
}

\section{Fields}{


\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} of predicates (functions
that return either \code{TRUE} or \code{FALSE}). These functions must have exactly
two arguments: \code{event} and \code{obj}. When the \code{filter()} method of the
Filterable is invoked on a LogEvent, that event will get passed to \code{event},
the Filterable will get passed to \code{obj}. This means that you can,
f.e. use \code{obj$threshold} to access the threshold of the Appender/Logger
that is using the filter.
If all of these functions evaluate to \code{TRUE} the LogEvent is passed on.
Since LogEvents have reference semantics, filters can also be abused to
modify them before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{

\describe{
\item{\code{append(event)}}{Tell the Appender to process a \link{LogEvent} \code{event}.
This method is usually not called by the user, but invoked by a
\link{Logger}
}
}


\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
}
}

\seealso{
Other abstract classes: \code{\link{AppenderDigest}},
  \code{\link{AppenderMemory}},
  \code{\link{AppenderTable}}, \code{\link{Filterable}}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderDigest}}, \code{\link{AppenderDt}},
  \code{\link{AppenderFile}}, \code{\link{AppenderGmail}},
  \code{\link{AppenderJson}}, \code{\link{AppenderMemory}},
  \code{\link{AppenderPushbullet}},
  \code{\link{AppenderRjdbc}},
  \code{\link{AppenderSendmail}},
  \code{\link{AppenderTable}}
}
\concept{Appenders}
\concept{abstract classes}
\keyword{internal}
