% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\docType{data}
\name{AppenderBuffer}
\alias{AppenderBuffer}
\title{AppenderBuffer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppenderBuffer
}
\description{
An Appender that Buffers LogEvents in-memory and and redirects them to other
appenders once certain conditions are met.
}
\section{Usage}{
\preformatted{x <- AppenderMemoryDt$new(threshold = NA_integer_,
  layout = LayoutFormat$new(fmt = "\%L [\%t] \%m", timestamp_fmt = "\%H:\%M:\%S",
  colors = getOption("yog.colors")), prototype = data.table::data.table(.id =
  NA_integer_, level = NA_integer_, timestamp = Sys.time(), caller =
  NA_character_, msg = NA_character_), buffer_size = 1e+05)

x$append(event)
x$show(n = 20, threshold = NA_integer_)
x$clone(deep = FALSE)

x$data
x$destination
}
}

\section{Creating a new AppenderBuffer}{


\describe{
\item{\code{buffer_size}}{\code{integer} scalar. Number of \link{LogEvents} to buffer}
\item{\code{should_flush}}{A \code{function} with a single argument \code{event}
(a \link{LogEvent}) that must only return either \code{TRUE} or \code{FALSE}. If the
function returns \code{TRUE}, flushing of the buffer is triggered. Defaults
to flushing if a \code{FATAL} event is registered
}
\item{\code{flush_on_exit}}{\code{TRUE} or \code{FALSE}: Whether the buffer should be
flushed when the Appender is garbage collected (f.e when you close R)}
\item{\code{flush_on_rotate}}{\code{TRUE} or \code{FALSE}: Whether the buffer should be
flushed when the Buffer is full (f.e when you close R). Setting this
to off can have slighly negative perfomance impacts.}
\item{\code{appenders}}{Like for a \link{Logger}. Buffered events will be passed on
to these Appenders once a flush is triggered}
}
}

\section{Methods}{


\describe{
\item{\code{flush()}}{Manually trigger a flush of the buffer}
\item{\code{add_appender(appender, name)}}{Add an Appender to this Appender.
see \link{Logger}}
\item{\code{remove_appender(pos)}}{Remove and Appender from this Appender.
see \link{Logger}}
}
}

\section{Creating a new Appender}{


\describe{
\item{threshold}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{layout}{A \link{Layout}. See examples.}
}
}

\examples{
# create a new logger with propagate = FALSE to prevent routing to the root
# logger. Please look at the section "Logger Hirarchies" in the package
# vignette for more info.
logger  <- Logger$new("testlogger", propagate = FALSE)

logger$add_appender(AppenderConsole$new())
logger$add_appender(AppenderConsole$new(
  layout = LayoutFormat$new("[\%t] \%c(): [\%n] \%m from user \%u", colors = getOption("yog.colors"))))

# Will output the message twice because we attached two console appenders
logger$warn("A test message")

}
\seealso{
\link{LayoutFormat}

Other Appenders: \code{\link{AppenderConsole}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}}, \code{\link{Appender}}
}
\concept{Appenders}
\keyword{datasets}
