% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderBuffer}
\alias{AppenderBuffer}
\title{Log to a Memory Buffer}
\description{
An Appender that Buffers LogEvents in-memory and and redirects them to other
Appenders once certain conditions are met.
}
\section{Usage}{
\preformatted{
x <- x$new(threshold = NA_integer_, appenders = NULL,
  should_flush = function(event) isTRUE(event[["level"]] <= 100),
  flush_on_exit = TRUE, flush_on_rotate = TRUE, layout = LayoutFormat$new(fmt =
  "\%L [\%t] \%m", timestamp_fmt = "\%H:\%M:\%S", colors = getOption("lgr.colors")),
  buffer_size = 1000)

x$append(x)
x$flush(x = private$.buffered_events)
x$set_appenders(x)
x$add_appender(appender, name = NULL)
x$remove_appender(pos)
x$set_buffer_size(x)
x$set_flush_on_exit(x)
x$set_flush_on_rotate(x)
x$set_should_flush(x)
x$show(threshold = NA_integer_, n = 20L)
x$append(event)
x$set_threshold(level)
x$set_layout(layout)
x$filter(event)
x$set_filters(filters)

x$appenders
x$buffer_size
x$buffered_events
x$data
x$destination
x$dt
x$flush_on_exit
x$flush_on_rotate
x$layout
x$threshold
x$filters

}
}

\section{Creating a Buffer Appender}{


The \link{Layout} for this Appender is used only to format console output of
its \code{$show()} method.
}

\section{Fields}{


\describe{
\item{\code{appenders}, \code{set_appenders()}}{Like for a \link{Logger}. Buffered events will be passed on
to these Appenders once a flush is triggered}
\item{\code{buffer_size, set_buffer_size(x)}}{\code{integer} scalar. Number of \link{LogEvents} to buffer}
\item{\code{flush_on_exit, set_flush_on_exit(x)}}{\code{TRUE} or \code{FALSE}: Whether the
buffer should be flushed when the Appender is garbage collected (f.e when
you close \R)}
\item{\code{flush_on_rotate, set_flush_on_rotate}}{\code{TRUE} or \code{FALSE}: Whether
the buffer should be flushed when the Buffer is full (f.e when you close
\R). Setting this to off can have slighly negative perfomance impacts.}
\item{\code{should_flush(event)}, \code{set_should_flush(x)}}{ A \code{function} with a
single argument \code{event} (a \link{LogEvent}) that must only return either \code{TRUE}
or \code{FALSE}. If the function returns \code{TRUE}, flushing of the buffer is
triggered. Defaults to flushing if a \code{FATAL} event is registered }
}



\describe{
\item{\code{dt}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}



\describe{
\item{\code{data}}{Get the log recorded by this \code{Appender} as a
\code{data.frame}}
}



\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} of predicates (functions
that return either \code{TRUE} or \code{FALSE}). If all of these functions evaluate
to \code{TRUE} the LogEvent is passed on. Since
LogEvents have reference semantics, filters can also be abused to modify
LogEvents before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{


\describe{
\item{\code{flush()}}{Manually trigger flushing}
\item{\code{add_appender(appender)}}{Add and Appender to this Appender.
see \link{Logger}}
\item{\code{remove_appender(pos)}}{Remove and Appender from this Appender.
see \link{Logger}}
}


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via this Appenders \link{Layout}}
}


\describe{
\item{\code{append(event)}}{Write the \link{LogEvent} \code{event} to a destination}
}


\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
}
}

\section{Comparison AppenderBuffer and AppenderDt}{


Both \link{AppenderBuffer} and \link{AppenderDt} do in memory buffering of events.
AppenderBuffer retains a copies of the events it processes and has the
abillity to pass the buffered events on to other Appenders. AppenderDt
converts the events to rows in a \code{data.table} and is a bit harder to
configure, but has much less overhead than AppenderBuffer.

Use AppenderDt if you want an in-memory log for interactive use, and
AppenderBuffer if you actually want to buffer events
}

\seealso{
\link{LayoutFormat}

Other Appenders: \code{\link{AppenderConsole}},
  \code{\link{AppenderDbi}}, \code{\link{AppenderDt}},
  \code{\link{AppenderFile}}, \code{\link{AppenderJson}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
