% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderBuffer}
\alias{AppenderBuffer}
\title{Log to a Memory Buffer}
\description{
An Appender that Buffers LogEvents in-memory and and redirects them to other
appenders once certain conditions are met.
}
\section{Usage}{
\preformatted{x <- AppenderBuffer$new(threshold = NA_integer_, appenders = NULL,
  should_flush = function(event) isTRUE(event[["level"]] <= 100),
  flush_on_exit = TRUE, flush_on_rotate = TRUE, layout = LayoutFormat$new(fmt =
  "\%L [\%t] \%m", timestamp_fmt = "\%H:\%M:\%S", colors = getOption("yog.colors")),
  buffer_size = 1000)

x$append(x)
x$flush(x = private$.buffered_events)
x$set_appenders(x)
x$add_appender(appender, name = NULL)
x$remove_appender(pos)
x$set_buffer_size(x)
x$set_flush_on_exit(x)
x$set_flush_on_rotate(x)
x$set_should_flush(x)
x$clone(deep = FALSE)

x$appenders
x$buffer_size
x$buffered_events
x$destination
x$flush_on_exit
x$flush_on_rotate
}
}

\section{Creating a New Appender}{


\describe{
\item{\code{buffer_size}}{\code{integer} scalar. Number of \link{LogEvents} to buffer}
\item{\code{should_flush}}{A \code{function} with a single argument \code{event}
(a \link{LogEvent}) that must only return either \code{TRUE} or \code{FALSE}. If the
function returns \code{TRUE}, flushing of the buffer is triggered. Defaults
to flushing if a \code{FATAL} event is registered
}
\item{\code{flush_on_exit}}{\code{TRUE} or \code{FALSE}: Whether the buffer should be
flushed when the Appender is garbage collected (f.e when you close R)}
\item{\code{flush_on_rotate}}{\code{TRUE} or \code{FALSE}: Whether the buffer should be
flushed when the Buffer is full (f.e when you close R). Setting this
to off can have slighly negative perfomance impacts.}
\item{\code{appenders}}{Like for a \link{Logger}. Buffered events will be passed on
to these Appenders once a flush is triggered}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{flush()}}{Manually trigger flushing}
\item{\code{add_appender(appender)}}{Add and Appender to this Appender.
see \link{Logger}}
\item{\code{remove_appender(pos)}}{Remove and Appender from this Appender.
see \link{Logger}}
\item{\code{appenders}, \code{set_appenders()}}{Get/set the \code{list} of \code{Appenders}
attached to this \code{Appender}}
\item{\code{buffer_size, set_buffer_size(x)}}{get/set the bufffer size}
\item{\code{flush_on_exit, set_flush_on_exit(x)}}{get/set whether the buffer
should be flushed when this \code{Appender} is garbage collected}
\item{\code{flush_on_rotate, set_flush_on_rotate}}{get/set whether the buffer
should be flushed once it is full}
\item{\code{should_flush(event)}, \code{set_should_flush(x)}}{Call or set the
function that determins whether an event should trigger flushing based
on its properties (e.g. its log level)
}
}



\describe{
\item{\code{append(event)}}{Write the \link{LogEvent} \code{event} to a destination}
\item{\code{threshold}, \code{set_threshold(level)}}{get/set the \code{Appender} threshold}
\item{\code{layout}, \code{set_layout(layout)}}{get/set the \code{Layout}}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}
}

\seealso{
\link{LayoutFormat}

Other Appenders: \code{\link{AppenderConsole}},
  \code{\link{AppenderDbi}}, \code{\link{AppenderFile}},
  \code{\link{AppenderJson}},
  \code{\link{AppenderMemoryDt}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
