% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderConsole}
\alias{AppenderConsole}
\title{AppenderConsole}
\description{
A simple Appender that outputs to the console. If you
have the package \strong{crayon} installed log levels will be coloured by default
(but you can modify this behaviour by passing a custom \link{Layout})
}
\section{Creating a new Appender}{


\describe{
\item{threshold}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{layout}{A \link{Layout}. See examples.}
}
}

\examples{
# create a new logger with propagate = FALSE to prevent routing to the root
# logger. Please look at the section "Logger Hirarchies" in the package
# vignette for more info.
logger  <- Logger$new("testlogger", propagate = FALSE)

logger$add_appender(AppenderConsole$new())
logger$add_appender(AppenderConsole$new(
  layout = LayoutFormat$new("[\%t] \%c(): [\%n] \%m from user \%u", colors = getOption("yog.colors"))))

# Will output the message twice because we attached two console appenders
logger$warn("A test message")

}
\seealso{
\link{LayoutFormat}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}}, \code{\link{Appender}}
}
\concept{Appenders}
