% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderDbi}
\alias{AppenderDbi}
\title{AppenderDbi}
\description{
Log to a database table with any DBI compatabile backend.
}
\section{Usage}{
\preformatted{x <- AppenderDbi$new(conn, table, field_types = c(level = "smallint", timestamp
  = "timestamp", caller = "varchar(1024)", msg = "varchar(1024)"),
  threshold = NA_integer_, layout = LayoutFormat$new(fmt = "\%L [\%t] \%m",
  timestamp_fmt = "\%H:\%M:\%OS3", colors = getOption("yog.colors", list())),
  close_on_exit = TRUE)

x$set_close_on_exit(x)
x$show(n = 20, threshold = NA_integer_)
x$append(event)
x$clone(deep = FALSE)

x$close_on_exit
x$conn
x$data
x$destination
}
}

\section{Creating a new Appender}{


\describe{
\item{conn}{a DBI connection}
\item{field_types}{a named \code{character} vector of field names and types.}
}
}

\section{Fields and Methods}{




\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via the defined \link{Layout}}
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}
}

\seealso{
Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}},
  \code{\link{AppenderRjdbc}}
}
\concept{Appenders}
