% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderDbi}
\alias{AppenderDbi}
\title{Log to Databases via DBI}
\description{
Log to a database table with any \strong{DBI} compatabile backend.
}
\section{Usage}{
\preformatted{x <- AppenderDbi$new(conn, table, threshold = NA_integer_,
  layout = select_dbi_layout(conn), close_on_exit = TRUE)

x$set_close_on_exit(x)
x$show(n = 20, threshold = NA_integer_)
x$append(event)
x$clone(deep = FALSE)

x$close_on_exit
x$conn
x$data
x$destination
x$table
}
}

\section{Creating a New Appender}{


\describe{
\item{conn}{a DBI connection}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{

\describe{
\item{\code{close_on_exit}, \code{set_close_on_exit()}}{\code{TRUE} or \code{FALSE}. Close the
Database connection when the Logger is removed?}
\item{\code{data}}{Querry the whole log from the Database and return it as a
\code{data.frame}}
\item{\code{conn}}{get the DBI connection object}
\item{\code{table}}{Name of the target database table}
}



\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via the defined \link{Layout}}
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}
}

\seealso{
Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}},
  \code{\link{AppenderFile}}, \code{\link{AppenderJson}},
  \code{\link{AppenderMemoryDt}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
