% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderDt}
\alias{AppenderDt}
\alias{yog_data}
\title{Log to an In-Memory Data.Table}
\description{
An Appender that outputs to an in-memory \code{data.table}. This requires that
you have the suggested package \strong{data.table} installed. This kind of
Appender is pretty useful for interactive use, and hast very little overhead.
}
\section{Usage}{
\preformatted{x <- AppenderDt$new(threshold = NA_integer_, layout = LayoutFormat$new(fmt =
  "\%L [\%t] \%m", timestamp_fmt = "\%H:\%M:\%S", colors = getOption("yog.colors")),
  prototype = data.table::data.table(.id = NA_integer_, level = NA_integer_,
  timestamp = Sys.time(), caller = NA_character_, msg = NA_character_),
  buffer_size = 1e+05)

x$append(event)
x$show(n = 20, threshold = NA_integer_)
x$clone(deep = FALSE)

x$data
x$destination
}
}

\section{Creating a New Appender}{


\describe{
\item{buffer_size}{\code{integer} scalar. Number of rows of the in-memory
\code{data.table}}
\item{prototype}{A prototype \code{data.table}. This is only necessary to set
manually if you use custom \link{LogEvents} (which is not yet supported by yog,
but planned for the future)}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via the defined \link{Layout}}
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}
}

\examples{
lg <- Logger$new(
  "test",
  appenders = list(memory = AppenderDt$new()),
  threshold = NA,
  parent = NULL  # to prevent routing to root logger for this example
)
lg$debug("test")
lg$error("test")

# Displaying the log
lg$appenders$memory$data
lg$appenders$memory$show()
show_log(target = lg$appenders$memory)

# If you pass a Logger to show_log, it looks for the first MemoryAppender it
# can find.
show_log(target = lg)
}
\seealso{
\link{LayoutFormat}, \link{simple_logging}, \link[data.table:data.table]{data.table::data.table}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderFile}}, \code{\link{AppenderJson}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
