% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderFile}
\alias{AppenderFile}
\title{AppenderFile}
\description{
A simple Appender that outputs to a file in the file system.
}
\section{Usage}{
\preformatted{x <- AppenderFile$new(file, threshold = NA_integer_,
  layout = LayoutFormat$new())

x$append(event)
x$set_file(file)
x$clone(deep = FALSE)

x$destination
x$file
}
}

\section{Creating a new AppenderFile}{


\describe{
\item{file}{\code{character} scalar. Path to the desired log file. If the file
does not exist it will be created}
}
}

\section{Fields}{


\describe{
\item{file}{\code{character} scalar. Path to the log file}
}



\describe{
\item{\code{threshold}}{An \code{integer}. Threshold of the current Appender
(i.e the maximum log level) that this Appender processes)}

\item{\code{layout}}{The \link{Layout} this Appender usses for formatting \link{LogEvent}s}

\item{\code{logger}}{The \link{Logger} this Appender is attached to}

\item{\code{destination}}{A \code{character} scalar giving the output destination
of the Appender as human readable print output.}
}
}

\section{Creating a new Appender}{


\describe{
\item{threshold}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{layout}{A \link{Layout}. See examples.}
}
}

\examples{
logger <- Logger$new("loggername")
default <- tempfile()
fancy <- tempfile()
json <- tempfile()

logger$add_appender(AppenderFile$new(default), "default")
logger$add_appender(
  AppenderFile$new(fancy, layout = LayoutFormat$new("[\%t] \%c(): \%L \%m from user \%u")), "fancy"
)
logger$add_appender(
  AppenderFile$new(json, layout = LayoutJson$new()), "json"
)

logger$info("A test message")

readLines(default)
readLines(fancy)
readLines(json)
}
\seealso{
\link{LayoutFormat}, \link{LayoutJson}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}},
  \code{\link{AppenderMemoryDt}}, \code{\link{Appender}}
}
\concept{Appenders}
