% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderFile}
\alias{AppenderFile}
\title{Log to a File}
\description{
A simple Appender that outputs to a file in the file system. \code{AppenderJson}
is just a shortcut for \code{AppenderFile} with \code{\link{LayoutJson}} as the default
\link{Layout}
}
\section{Usage}{
\preformatted{x <- AppenderFile$new(file, threshold = NA_integer_,
  layout = LayoutFormat$new())

x$append(event)
x$set_file(file)
x$clone(deep = FALSE)

x$destination
x$file
}
}

\section{Creating a New Appender}{


\describe{
\item{file}{\code{character} scalar. Path to the desired log file. If the file
does not exist it will be created}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{file, set_file(file)}}{get/set the log file}
}



\describe{
\item{\code{append(event)}}{Write the \link{LogEvent} \code{event} to a destination}
\item{\code{threshold}, \code{set_threshold(level)}}{get/set the \code{Appender} threshold}
\item{\code{layout}, \code{set_layout(layout)}}{get/set the \code{Layout}}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}
}

\examples{
logger <- Logger$new("loggername")
default <- tempfile()
fancy <- tempfile()
json <- tempfile()

logger$add_appender(AppenderFile$new(default), "default")
logger$add_appender(
  AppenderFile$new(fancy, layout = LayoutFormat$new("[\%t] \%c(): \%L \%m from user \%u")), "fancy"
)
logger$add_appender(
  AppenderFile$new(json, layout = LayoutJson$new()), "json"
)

logger$info("A test message")

readLines(default)
readLines(fancy)
readLines(json)
}
\seealso{
\link{LayoutFormat}, \link{LayoutJson}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderJson}},
  \code{\link{AppenderMemoryDt}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
