% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderFile}
\alias{AppenderFile}
\title{AppenderFile}
\description{
A simple Appender that outputs to a file in the file system.
}
\section{Creating a new AppenderFile}{


\describe{
\item{file}{\code{character} scalar. Path to the desired log file. If the file
does not exist it will be created}
}
}

\section{Creating a new Appender}{


\describe{
\item{threshold}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{layout}{A \link{Layout}. See examples.}
}
}

\section{Usage}{
\preformatted{Appender$new(layout = Layout$new(), threshold = NA_integer_)
# public methods
    app$append(event)
    app$set_threshold(level)
    app$set_logger(logger)
    app$set_layout(layout)
# public fields
    app$
# active bindings
    app$threshold
    app$layout
    app$logger
    app$destination
}
}

\examples{
logger <- Logger$new("loggername")
default <- tempfile()
fancy <- tempfile()
json <- tempfile()

logger$add_appender(AppenderFile$new(default), "default")
logger$add_appender(
  AppenderFile$new(fancy, layout = LayoutFormat$new("[\%t] \%c(): \%L \%m from user \%u")), "fancy"
)
logger$add_appender(
  AppenderFile$new(json, layout = LayoutJson$new()), "json"
)

logger$info("A test message")

readLines(default)
readLines(fancy)
readLines(json)
}
\seealso{
\link{LayoutFormat}, \link{LayoutJson}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}},
  \code{\link{AppenderMemoryDt}}, \code{\link{Appender}}
}
\concept{Appenders}
