% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderFileRotatingDate}
\alias{AppenderFileRotatingDate}
\alias{AppenderFileRotatingDateTime}
\title{Log to a rotating file}
\description{
An extension of \link{AppenderFile} that rotates logfiles based on certain
conditions. Please refert to the documentation of \code{\link[rotor:rotate]{rotor::rotate()}} for
the meanings of the extra arguments
}
\section{Usage}{
\preformatted{
x <- AppenderFileRotatingDate$new(file, threshold = NA_integer_, layout =
  LayoutFormat$new(), filters = NULL, age = NULL, timestamp_fmt = "\%Y-\%m-\%d",
  size = 1, max_backups = Inf, compression = FALSE, overwrite = FALSE,
  create_file = TRUE)

x$rotate(dry_run = getOption("rotor.dry_run", FALSE), verbose = getOption("rotor.dry_run", dry_run), now = Sys.Date())
x$set_timestamp_fmt(x)
x$clone(deep = FALSE)
x$prune(max_backups = self$max_backups)
x$set_age(x)
x$set_overwrite(x)
x$rotate(dry_run = getOption("rotor.dry_run", FALSE), verbose = getOption("rotor.dry_run", dry_run))
x$set_size(x)
x$set_max_backups(x)
x$set_compression(x)
x$set_create_file(x)
x$append(event)
x$set_file(file)
x$set_threshold(level)
x$set_layout(layout)
x$filter(event)
x$add_filter(filter, name = NULL)
x$remove_filter(pos)
x$set_filters(filters)

x$age
x$backups
x$compression
x$create_file
x$max_backups
x$overwrite
x$size
x$timestamp_fmt
x$destination
x$file
x$layout
x$threshold
x$filters

}
}

\section{Fields}{


\describe{
\item{\code{age}, \code{timestamp_fmt}, \code{overwrite}, \code{compression}, \code{size}}{
see \code{\link[rotor:rotate]{rotor::rotate()}}}
}



\describe{
\item{\code{file}, \code{set_file(file)}}{\code{character} scalar. Path to the desired log
file. If the file does not exist it will be created}
}



\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} that may contain
\code{functions} or any \R object with a \code{filter()} method. These functions
must have exactly one argument: \code{event} which will get passed the
LogEvent when the Filterable's \code{filter()} method is invoked.
If all of these functions evaluate to \code{TRUE} the LogEvent is passed on.
Since LogEvents have reference semantics, filters can also be abused to
modify them before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Creating a New Appender}{


New Appenders are instantiated with \code{<AppenderSubclass>$new()}. For the
arguments to \code{new()} please refer to the section \emph{Fields}. You can also
modify those fields after the Appender has been created with setters in the
form of \code{appender$set_<fieldname>(value)}
}

\section{Methods}{

\describe{
\item{\code{append(event)}}{Tell the Appender to process a \link{LogEvent} \code{event}.
This method is usually not called by the user, but invoked by a
\link{Logger}
}
}


\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
\item{\code{add_filter(filter, name = NULL)}, \code{remove_filter(pos)}}{
Add or remove a filter. When adding a filter an optional \code{name} can
be specified. \code{remove_filter()} can remove by position or name (if one
was specified)
}
}
}

\seealso{
\link{LayoutFormat}, \link{LayoutJson}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderFileRotatingDateTime}},
  \code{\link{AppenderFileRotating}},
  \code{\link{AppenderFile}}, \code{\link{AppenderGmail}},
  \code{\link{AppenderJson}},
  \code{\link{AppenderPushbullet}},
  \code{\link{AppenderRjdbc}},
  \code{\link{AppenderSendmail}},
  \code{\link{AppenderTable}}, \code{\link{Appender}}
}
\concept{Appenders}
