% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderJson}
\alias{AppenderJson}
\title{Log to a JSON File}
\description{
\code{AppenderJson} is a shortcut for \code{AppenderFile} with \code{\link{LayoutJson}}, but
comes with an extra method \code{show()} and an extra active field \code{data} to
access the underlying file.
}
\section{Usage}{
\preformatted{x <- AppenderFile$new(file, threshold = NA_integer_,
  layout = LayoutFormat$new())

x$append(event)
x$set_file(file)
x$clone(deep = FALSE)

x$destination
x$file
}
}

\section{Creating a New Appender}{




\describe{
\item{file}{\code{character} scalar. Path to the desired log file. If the file
does not exist it will be created}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated as in the source
JSON file}
\item{\code{data}}{The log recorded by this \code{Appender} as a \code{data.frame}}
}



\describe{
\item{\code{file, set_file(file)}}{get/set the log file}
}



\describe{
\item{\code{append(event)}}{Write the \link{LogEvent} \code{event} to a destination}
\item{\code{threshold}, \code{set_threshold(level)}}{get/set the \code{Appender} threshold}
\item{\code{layout}, \code{set_layout(layout)}}{get/set the \code{Layout}}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filter(event)}}{Check if an event passes all filters}
\item{\code{filters},\code{set_filters(filters)}}{get/set a list of \code{filters}. A
filter is a function that takes exactly two named arguments: the LogEvent
\code{event} and the Filterable \code{obj} (usually a Logger or Appender)}
}
}

\examples{
tf <- tempfile()
l <- Logger$new("testlogger", appenders = AppenderJson$new(tf), propagate = FALSE)

l$info("A test message")
l$info("A test message \%s strings", "with format strings", and = "custom_fields")

l$appenders[[1]]$show()
l$appenders[[1]]$data

}
\seealso{
\link{LayoutFormat}, \link{LayoutJson}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderFile}},
  \code{\link{AppenderMemoryDt}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
