% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderPushbullet}
\alias{AppenderPushbullet}
\title{Log to Pushbullet}
\description{
Send push notifications to \href{https://www.pushbullet.com/}{pushbullet}. This
Appender keeps an in-memory buffer like \link{AppenderBuffer}. If the buffer is
flushed, usually becuase an event of specified magnitutde is encountered, all
buffered events are concatenated to a single message that is sent to
\code{\link[RPushbullet:pbPost]{RPushbullet::pbPost()}}. The default behaviour is to push the last 7 log
events in case a \code{fatal} event is encountered.
}
\section{Usage}{
\preformatted{
x <- x$new(threshold = NA_integer_, flush_level = "fatal",
  layout = LayoutFormat$new(fmt = "\%K \%t> \%m \%f", timestamp_fmt = "\%H:\%M:\%S",
  colors = NULL), buffer_size = 6, apikey = NULL)

x$flush()
x$set_apikey(x)
x$append(event)
x$set_buffer_size(x)
x$set_should_flush(x)
x$set_flush_level(level)
x$show(threshold = NA_integer_, n = 20L)
x$set_threshold(level)
x$set_layout(layout)
x$filter(event)
x$set_filters(filters)

x$apikey
x$buffer_size
x$buffered_events
x$data
x$dt
x$destination
x$layout
x$threshold
x$filters

}
}

\section{Fields}{


\describe{
\item{\code{apikey}, \code{set_apikey()}}{See \code{\link[RPushbullet:pbPost]{RPushbullet::pbPost()}}}
}



\describe{
\item{\code{buffer_size, set_buffer_size(x)}}{\code{integer} scalar. Number of \link{LogEvents} to buffer}
\item{\code{flush_on_exit, set_flush_on_exit(x)}}{\code{TRUE} or \code{FALSE}: Whether the
buffer should be flushed when the Appender is garbage collected (f.e when
you close \R)}
\item{\code{flush_on_rotate, set_flush_on_rotate}}{\code{TRUE} or \code{FALSE}: Whether
the buffer should be flushed when the Buffer is full (f.e when you close
\R). Setting this to off can have slighly negative perfomance impacts.}
\item{\code{flush_level}, \code{set_flush_level()}}{\code{integer} or \code{character}
\link[=log_level]{log level}. Minimum event level that will trigger flushing of
the buffer. This behaviour is implemented through \code{should_flush()},
and you can modify that function for different behaviour.
}
\item{\code{should_flush(event, obj)}, \code{set_should_flush(x)}}{
A function with exactly two arguments: \code{event} and \code{obj}. The LogEvent
will be passed to \code{event}, the Appender will get passed to \code{obj}.
This means that you can f.e. use \code{obj$threshold} to access the threshold
of the Appender. If the function returns \code{TRUE}, flushing of the buffer
is triggered. Defaults to flushing if a \code{FATAL} event is registered }
}



\describe{
\item{\code{dt}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}



\describe{
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.frame}}
}



\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} of predicates (functions
that return either \code{TRUE} or \code{FALSE}). These functions must have exactly
two arguments: \code{event} and \code{obj}. When the \code{filter()} method of the
Filterable is invoked on a LogEvent, that event will get passed to \code{event},
the Filterable will get passed to \code{obj}. This means that you can,
f.e. use \code{obj$threshold} to access the threshold of the Appender/Logger
that is using the filter.
If all of these functions evaluate to \code{TRUE} the LogEvent is passed on.
Since LogEvents have reference semantics, filters can also be abused to
modify them before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{




\describe{
\item{\code{flush()}}{Manually trigger flushing of the buffer}
}


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via this Appenders \link{Layout}}
}


\describe{
\item{\code{append(event)}}{Tell the Appender to process a \link{LogEvent} \code{event}.
This method is usually not called by the user, but invoked by a
\link{Logger}
}
}


\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
}
}

\seealso{
\link{LayoutFormat}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderDt}}, \code{\link{AppenderFile}},
  \code{\link{AppenderJson}}, \code{\link{AppenderMemory}},
  \code{\link{AppenderRjdbc}}, \code{\link{AppenderTable}},
  \code{\link{Appender}}
}
\concept{Appenders}
