% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderRjdbc}
\alias{AppenderRjdbc}
\title{Log to Databases via RJDBC}
\description{
Log to a database table with the \strong{RJDBC} package. \strong{RJDBC} is only
somewhat  \strong{DBI} compliant and does not work with \link{AppenderDbi}. I
personally do not recommend using \strong{RJDBC} if it can be avoided.
}
\section{Usage}{
\preformatted{x <- AppenderRjdbc$new(conn, table, threshold = NA_integer_,
  layout = select_dbi_layout(conn), close_on_exit = TRUE)

x$append(event)
x$clone(deep = FALSE)

x$
}
}

\section{Creating a New Appender}{


\describe{
\item{conn}{an RJDBC connection}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{



\describe{
\item{\code{close_on_exit}, \code{set_close_on_exit()}}{...}
\item{\code{data}}{...}
\item{\code{conn}}{get the DBI connection object}
\item{\code{table}}{Name of the target database table}
}



\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via the defined \link{Layout}}
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}
}

\seealso{
\link{LayoutFormat}, \link{simple_logging}, \link[data.table:data.table]{data.table::data.table}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderFile}}, \code{\link{AppenderJson}},
  \code{\link{AppenderMemoryDt}},
  \code{\link{AppenderTable}}, \code{\link{Appender}}
}
\concept{Appenders}
