% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderSyslog}
\alias{AppenderSyslog}
\title{Log to the POSIX System Log}
\description{
An Appender that writes to the syslog on supported POSIX platforms. Requires
the \pkg{rsyslog} package.
}
\section{Fields}{


\describe{
\item{\code{identifier}}{\code{character} scalar. A string identifying the process;
if \code{NULL} defaults to the logger name}
\item{\code{syslog_levels}}{
\itemize{
\item a named \code{character} vector mapping whose names are log levels as
understood by \code{\link[rsyslog:syslog]{rsyslog::syslog()}} and whose values are
\link[=log_levels]{lgr log levels} (either \code{character} or \code{numeric})
\item a \code{function} that takes a vector of lgr log levels as input and returns a
\code{character} vector of log levels for \code{\link[rsyslog:syslog]{rsyslog::syslog()}}.
}
}
}
}

\examples{
if (requireNamespace("rsyslog", quietly = TRUE)) {
  lg <- get_logger("rsyslog/test")
  lg$add_appender(AppenderSyslog$new(), "syslog")
  lg$info("A test message")

  if (Sys.info()[["sysname"]] == "Linux"){
    system("journalctl -t 'rsyslog/test'")
  }

  invisible(lg$config(NULL))  # cleanup
}
}
\seealso{
\link{LayoutFormat}, \link{LayoutJson}

Other Appenders: 
\code{\link{AppenderBuffer}},
\code{\link{AppenderConsole}},
\code{\link{AppenderFile}},
\code{\link{AppenderJson}},
\code{\link{AppenderTable}},
\code{\link{Appender}}
}
\concept{Appenders}
\section{Super classes}{
\code{\link[lgr:Filterable]{lgr::Filterable}} -> \code{\link[lgr:Appender]{lgr::Appender}} -> \code{AppenderSyslog}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AppenderSyslog$new()}}
\item \href{#method-append}{\code{AppenderSyslog$append()}}
\item \href{#method-set_syslog_levels}{\code{AppenderSyslog$set_syslog_levels()}}
\item \href{#method-set_identifier}{\code{AppenderSyslog$set_identifier()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="add_filter">}\href{../../lgr/html/Filterable.html#method-add_filter}{\code{lgr::Filterable$add_filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="filter">}\href{../../lgr/html/Filterable.html#method-filter}{\code{lgr::Filterable$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="remove_filter">}\href{../../lgr/html/Filterable.html#method-remove_filter}{\code{lgr::Filterable$remove_filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Filterable" data-id="set_filters">}\href{../../lgr/html/Filterable.html#method-set_filters}{\code{lgr::Filterable$set_filters()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Appender" data-id="format">}\href{../../lgr/html/Appender.html#method-format}{\code{lgr::Appender$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Appender" data-id="set_layout">}\href{../../lgr/html/Appender.html#method-set_layout}{\code{lgr::Appender$set_layout()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lgr" data-topic="Appender" data-id="set_threshold">}\href{../../lgr/html/Appender.html#method-set_threshold}{\code{lgr::Appender$set_threshold()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppenderSyslog$new(
  identifier = NULL,
  threshold = NA_integer_,
  layout = LayoutFormat$new("\%m"),
  filters = NULL,
  syslog_levels = c(CRITICAL = "fatal", ERR = "error", WARNING = "warn", INFO = "info",
    DEBUG = "debug", DEBUG = "trace")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppenderSyslog$append(event)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_syslog_levels"></a>}}
\subsection{Method \code{set_syslog_levels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppenderSyslog$set_syslog_levels(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_identifier"></a>}}
\subsection{Method \code{set_identifier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppenderSyslog$set_identifier(x)}\if{html}{\out{</div>}}
}

}
}
