% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderTable}
\alias{AppenderTable}
\title{Abstract class for logging to tabular structures}
\description{
This is an \strong{abstract class} and as such cannot be instantiated directly.
It is exported for package developers that want to extend it

AppenderTable is extended by Appenders that write to a data source that
can be interpreted as tables, (usually a \code{data.frame}). Examples are
\link[lgrExtra:AppenderDbi]{lgrExtra::AppenderDbi}, \link[lgrExtra:AppenderRjdbc]{lgrExtra::AppenderRjdbc} and \link[lgrExtra:AppenderDt]{lgrExtra::AppenderDt}.
}
\section{Fields}{


\describe{
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.frame}}
}



\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}
}

\section{Methods}{


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a
log level bellow \code{threshold}.}
}


\describe{
\item{\code{append(event)}}{Tell the Appender to process a \link{LogEvent} \code{event}.
This method is usually not called by the user, but invoked by a
\link{Logger}
}
}
}

\seealso{
Other abstract classes: 
\code{\link{AppenderMemory}},
\code{\link{Filterable}}

Other Appenders: 
\code{\link{AppenderBuffer}},
\code{\link{AppenderConsole}},
\code{\link{AppenderFileRotating}},
\code{\link{AppenderFile}},
\code{\link{AppenderJson}},
\code{\link{AppenderSyslog}},
\code{\link{Appender}}
}
\concept{Appenders}
\concept{abstract classes}
