% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\docType{data}
\name{AppenderTable}
\alias{AppenderTable}
\title{AppenderTable}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppenderTable
}
\description{
Log to a database table with the RJDBC package.
}
\section{Creating a new Appender}{


\describe{
\item{conn}{an RJDBC connection}
\item{col_types}{a named \code{character} vector of field names and types.}
}



\describe{
\item{\code{threshold}}{\code{character} or \code{integer} scalar. The minimum log level
that triggers this logger. See \link{log_levels}}
\item{\code{layout}}{a \code{Layout} that will be used for formatting the \code{LogEvents}
passed to this Appender
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a log level
bellow \code{threshold}. The log entries will be formated for console output
via the defined \link{Layout}}
\item{\code{data}}{Get the log recorded by this \code{Appender} as a \code{data.table}
with a maximum of \code{buffer_size} rows}
}
}

\seealso{
\link{LayoutFormat}, \link{simple_logging}, \link[data.table:data.table]{data.table::data.table}

Other internal Appenders: \code{\link{Appender}}
}
\concept{internal Appenders}
\keyword{datasets}
