% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appender.R
\name{AppenderTable}
\alias{AppenderTable}
\title{Abstract Class for Logging to Tabular Structures}
\description{
Abstract classes are exported for developers that want to extend them, they
are not useful to casual users. \link{AppenderDbi}, \link{AppenderRjdbc} and
\link{AppenderDt} are derived from AppenderTabel.
}
\section{Fields}{


\describe{
\item{\code{data}}{Get the log recorded by this \code{Appender} as a
\code{data.frame}}
}



\describe{
\item{\code{threshold}, \code{set_threshold(level)}}{\code{character} or \code{integer} scalar.
The minimum log level that triggers this logger. See \link{log_levels}}
\item{\code{layout}, \code{set_layout(layout)}}{a \code{Layout} that will be used for
formatting the \code{LogEvents} passed to this Appender}
\item{\code{destination}}{The output destination of the \code{Appender} in
human-readable form (mainly for print output)}
}



\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} of predicates (functions
that return either \code{TRUE} or \code{FALSE}). If all of these functions evaluate
to \code{TRUE} the LogEvent is passed on. Since
LogEvents have reference semantics, filters can also be abused to modify
LogEvents before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{


\describe{
\item{\code{show(n, threshold)}}{Show the last \code{n} log entries with a
log level bellow \code{threshold}.}
}


\describe{
\item{\code{append(event)}}{Write the \link{LogEvent} \code{event} to a destination}
}


\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
}
}

\seealso{
\link{LayoutTable}

Other Appenders: \code{\link{AppenderBuffer}},
  \code{\link{AppenderConsole}}, \code{\link{AppenderDbi}},
  \code{\link{AppenderDt}}, \code{\link{AppenderFile}},
  \code{\link{AppenderJson}}, \code{\link{AppenderRjdbc}},
  \code{\link{Appender}}
}
\concept{Appenders}
