% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filterable.R
\docType{data}
\name{Filterable}
\alias{Filterable}
\title{Abstract Class for Filterables}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Filterable
}
\description{
Superclass for classes that have a \code{filter()} method such as Appenders and
Loggers. This class is only exported for package developers that want to
extend it.
}
\section{Fields and Methods}{


\describe{
\item{\code{filter(event)}}{Check if an event passes all filters}
\item{\code{filters},\code{set_filters(filters)}}{get/set a list of \code{filters}. A
filter is a function that takes exactly two named arguments: the LogEvent
\code{event} and the Filterable \code{obj} (usually a Logger or Appender)}
}
}

\keyword{datasets}
\keyword{internal}
