% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filterable.R
\docType{data}
\name{Filterable}
\alias{Filterable}
\title{Abstract Class for Filterables}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Filterable
}
\description{
\strong{Abstract classes} are exported for package developers that want to extend
them, they cannot be instantiated directly.

Superclass for classes that have a \code{filter()} method such as \link{Appenders} and
\link{Loggers}. This class is only exported for package developers that want to
extend it.
}
\section{Fields}{


\describe{
\item{\code{filters}, \code{set_filters(filters)}}{a \code{list} of predicates (functions
that return either \code{TRUE} or \code{FALSE}). These functions must have exactly
two arguments: \code{event} and \code{obj}. When the \code{filter()} method of the
Filterable is invoked on a LogEvent, that event will get passed to \code{event},
the Filterable will get passed to \code{obj}. This means that you can,
f.e. use \code{obj$threshold} to access the threshold of the Appender/Logger
that is using the filter.
If all of these functions evaluate to \code{TRUE} the LogEvent is passed on.
Since LogEvents have reference semantics, filters can also be abused to
modify them before they are passed on. Look at the source code of
\code{\link[=with_log_level]{with_log_level()}} or \code{\link[=with_log_value]{with_log_value()}} for examples.
}
}
}

\section{Methods}{

\describe{
\item{\code{filter(event)}}{Determine whether the LogEvent \code{x} should be passed
on to Appenders (\code{TRUE}) or not (\code{FALSE}). See also the active binding
\code{filters}}
}
}

\seealso{
Other abstract classes: \code{\link{AppenderDigest}},
  \code{\link{AppenderMail}}, \code{\link{AppenderMemory}},
  \code{\link{AppenderTable}}
}
\concept{abstract classes}
\keyword{datasets}
\keyword{internal}
