% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{is_filter}
\alias{is_filter}
\title{Check if an R Object is a Filter}
\usage{
is_filter(x)
}
\arguments{
\item{x}{any \R Object}
}
\description{
Check if an R Object is a Filter
}
\seealso{
\link{EventFilter}
}
