% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{Layout}
\alias{Layout}
\alias{Layouts}
\title{Abstract Class for Layouts}
\description{
Abstract classes are exported for developers that want to extend them, they
are not useful to casual users.  Layouts get a \link{LogEvent} passed down from an
\link{Appender}, and format them for output. How this formatting works exactly
varries widely. For example for file or console output the log event is
usually formatted into a single character line.
}
\section{Fields and Methods}{


\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\seealso{
Other Layouts: \code{\link{LayoutDbi}},
  \code{\link{LayoutFormat}}, \code{\link{LayoutGlue}},
  \code{\link{LayoutJson}}, \code{\link{LayoutTable}}
}
\concept{Layouts}
\keyword{internal}
