% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{LayoutDbi}
\alias{LayoutDbi}
\alias{LayoutSqlite}
\alias{select_dbi_layout}
\title{LayoutDbi}
\description{
Format a \link{LogEvent} as data.frame for inserting into a Database.
}
\section{Usage}{
\preformatted{x <- LayoutDbi$new(event_vals = c("level", "timestamp", "caller", "msg"),
  logger_vals = character(), col_types = NULL)

x$format_event(event)
x$set_col_types(x)
x$sql_create_table(table)
x$clone(deep = FALSE)

x$col_names
x$col_types
}
}

\section{Creating a New Layout}{

\describe{
\item{\code{col_types}}{a named \code{character} vector of column types supported by
the target database. Must include all columns described in \code{event_vals},
\code{logger_vals}. If a new database table is created by
LayoutDbi its column order will correspond to the order of \code{col_types}.
}
}



\describe{
\item{\code{event_vals}}{Names of the fields of the \link{LogEvent}
to include in the output object.
}

\item{\code{logger_vals}}{Names of the fields of the \link{Logger} that produced the
\link{LogEvent} to include in the output object.
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{col_types}, \code{set_col_types()}}{Get/set the col_types of this \code{Layout}}
\item{\code{col_names}}{Convenience method to get the names of the \code{col_types}
vector}
}



\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\section{Database Specific Layouts}{


Different databases have different data types and features. Currently the
following \code{LayoutDBI} subclasses exist that deal with specific databases,
but this list is expected to grow as yog matures:
\itemize{
\item \code{LayoutSQLite}: Needs its own Layout because SQLite does not support
\code{timestamps}
\item \code{LayoutDBI}: For all other datbases
}

The utility function \code{select_dbi_layout()} returns the appropriate
Layout for a DBI connection.
}

\examples{
# setup a dummy LogEvent
event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message"
)

# defaults
lo <- LayoutDbi$new()
lo$format_event(event)

# SQLite does not support timestamps so LayoutSqlite converts them to text
lo <- LayoutSqlite$new()
str(lo$format_event(event))

# advanced example that supports a custom_field:
lo <- LayoutDbi$new(
  event_vals = c("level", "timestamp", "msg", "custom_field"),
  logger_vals = c(logger_name = "name", "user"),
  col_types =  c(
    timestamp = "timestamp",
    user = "varchar(256)",
    logger_name = "varchar(256)",
    level = "smallint",
    msg = "varchar(2048)",
    custom_field = "integer"
  )
)

event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message",
  custom_field = "blubb"
)

lo$format_event(event)
#'
}
\seealso{
\link[DBI:DBI]{DBI::DBI}

Other Layouts: \code{\link{LayoutFormat}},
  \code{\link{LayoutJson}}, \code{\link{LayoutTable}},
  \code{\link{Layout}}
}
\concept{Layouts}
\concept{database layouts}
