% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{LayoutDbi}
\alias{LayoutDbi}
\alias{LayoutSqlite}
\alias{select_dbi_layout}
\title{Format Log Events for Output to Databases}
\description{
Format a \link{LogEvent} as \code{data.frame} for inserting into a relational database.
}
\section{Usage}{
\preformatted{
x <- x$new(event_values = c("level", "timestamp", "caller", "msg"),
  col_types = NULL)

x$format_event(event)
x$set_col_types(x)
x$sql_create_table(table)
x$clone(deep = FALSE)
x$set_event_values(x)
x$set_val_names(x)

x$col_names
x$col_types
x$event_values
x$val_names

}
}

\section{Creating a New Layout}{

\describe{
\item{\code{col_types}}{a named \code{character} vector of column types supported by
the target database. This information is used by \link{AppenderDbi}
or similar Appenders to create a new database table, either on
instantion of the Appender or on writing of the first LogEvent.
If not null, it must include \emph{all} columns described in \code{event_values}, and
it will decide the column order of the \code{$format_event()} output.
}
}



\describe{
\item{\code{event_values}}{Names of the fields of the \link{LogEvent}
to include in the output object.
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{col_types}, \code{set_col_types()}}{Get/set the col_types of this \code{Layout}}
\item{\code{col_names}}{Convenience method to get the names of the \code{col_types}
vector}
}



\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\section{Database Specific Layouts}{


Different databases have different data types and features. Currently the
following \code{LayoutDBI} subclasses exist that deal with specific databases,
but this list is expected to grow as yog matures:
\itemize{
\item \code{LayoutSQLite}: Needs its own Layout because SQLite does not support
\code{timestamps}
\item \code{LayoutDBI}: For all other datbases
}

The utility function \code{select_dbi_layout()} returns the appropriate
Layout for a DBI connection.
}

\examples{
# setup a dummy LogEvent
event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message"
)

# defaults
lo <- LayoutDbi$new()
lo$format_event(event)

# SQLite does not support timestamps so LayoutSqlite converts them to text
lo <- LayoutSqlite$new()
str(lo$format_event(event))

# advanced example that supports a custom_field:
lo <- LayoutDbi$new(
  event_values = c("level", "timestamp", "msg", "custom_field"),
  col_types =  c(
    timestamp = "timestamp",
    level = "smallint",
    msg = "varchar(2048)",
    custom_field = "integer"
  )
)

event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message",
  custom_field = "blubb"
)

lo$format_event(event)
#'
}
\seealso{
\link[DBI:DBI]{DBI::DBI}

Other Layouts: \code{\link{LayoutFormat}},
  \code{\link{LayoutJson}}, \code{\link{LayoutTable}},
  \code{\link{Layout}}
}
\concept{Layouts}
\concept{database layouts}
