% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{LayoutDbi}
\alias{LayoutDbi}
\title{LayoutDbi}
\description{
Format a \link{LogEvent} as data.frame for inserting into a Database.
}
\section{Usage}{
\preformatted{x <- LayoutDbi$new(event_vals = c("level", "timestamp", "caller", "msg"),
  logger_vals = NULL, other_vals = NULL, col_types = NULL)

x$format_event(event)
x$set_col_types(x)
x$clone(deep = FALSE)

x$col_names
x$col_types
}
}

\section{Creating a New Layout}{

\describe{
\item{\code{col_types}}{a named \code{character} vector of column types supported by
the target database. Must include all columns described in \code{event_vals},
\code{logger_vals} and \code{other_vals}. If a new database table is created by
LayoutDbi its column order will correspond to the order of \code{col_types}.
}
}



\describe{
\item{\code{event_vals}}{Names of the fields of the \link{LogEvent}
to include in the output object.
}

\item{\code{logger_vals}}{Names of the fields of the \link{Logger} that produced the
\link{LogEvent} to include in the output object.
}

\item{\code{other_vals}}{A named \code{list} of any kind of \R values. Functions in
this `list`` will be executed with no arguments and their results will
be included in the results object (see examples)
}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{col_types}, \code{set_col_types()}}{Get/set the col_types of this \code{Layout}}
\item{\code{col_names}}{Convenience method to get the names of the \code{col_types}
vector}
}



\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\examples{
# setup a dummy LogEvent
event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message"
)

# defaults
lo <- LayoutDbi$new()
lo$format_event(event)


# advanced example:
lo <- LayoutDbi$new(
  event_vals = c("level", "timestamp", "msg"),
  logger_vals = "user",
  other_vals = list(pid = Sys.getpid, teststring = "blah"),
  col_types =  c(
    timestamp = "timestamp",
    level = "smallint",
    msg = "varchar(1024)",
    user = "varchar(256)",
    pid = "integer",
    teststring = "varchar(256)"
  )
)
lo$format_event(event)

}
\seealso{
\code{\link[=read_json_lines]{read_json_lines()}}, \url{http://jsonlines.org/}

Other Layouts: \code{\link{LayoutFormat}},
  \code{\link{LayoutJson}}, \code{\link{LayoutTable}},
  \code{\link{Layout}}
}
\concept{Layouts}
\concept{database layouts}
