% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{LayoutFormat}
\alias{LayoutFormat}
\title{LayoutFormat}
\description{
Format a \link{LogEvent} as human readable text using \code{\link[=format.LogEvent]{format.LogEvent()}}
}
\section{Usage}{
\preformatted{x <- LayoutFormat$new(fmt = "\%L [\%t] \%m", timestamp_fmt = "\%Y-\%m-\%d
  \%H:\%M:\%OS3", colors = NULL, pad_levels = "right")

x$format_event(event)
x$set_fmt(x)
x$set_timestamp_fmt(x)
x$set_colors(x)
x$set_pad_levels(x)
x$clone(deep = FALSE)

x$colors
x$fmt
x$pad_levels
x$timestamp_fmt
}
}

\section{Creating a New LayoutFormat}{


LayoutFormat passes it fields as arguments to \code{\link[=format.LogEvent]{format.LogEvent()}}.

\describe{
\item{\code{fmt}}{see \code{\link[=format.LogEvent]{format.LogEvent()}}}
\item{\code{timestamp_fmt}}{see \code{\link[base:format.POSIXct]{base::format.POSIXct()}}}
\item{\code{colors}}{see \code{\link[=format.LogEvent]{format.LogEvent()}}}
\item{\code{pad_levels}}{see \code{\link[=format.LogEvent]{format.LogEvent()}}}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\examples{

# setup a dummy LogEvent
event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message"
)
lo <- LayoutFormat$new()
lo$format_event(event)

}
\seealso{
Other Layouts: \code{\link{LayoutDbi}},
  \code{\link{LayoutJson}}, \code{\link{LayoutTable}},
  \code{\link{Layout}}
}
\concept{Layouts}
