% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{LayoutGlue}
\alias{LayoutGlue}
\title{Format Log Events as Text with Glue}
\description{
Format a \link{LogEvent} as human readable text using \link[glue:glue]{glue::glue}. The function
is evaluated in an environment in which it has access to all elements of
the \link{LogEvent} (see examples)
}
\section{Usage}{
\preformatted{
x <- x$new(fmt = "{pad_right(colorize_levels(toupper(level_name)), 5)}
  [{timestamp}] msg", colors = NULL)

x$format_event(event)
x$set_fmt(x)
x$set_colors(x)
x$clone(deep = FALSE)

x$fmt

}
}

\section{Creating a New Layout}{


\describe{
\item{\code{fmt}}{see \code{\link[glue:glue]{glue::glue()}}}
}
}

\section{Fields and Methods}{


\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\examples{
lg <- Logger$new("testlogger", appenders = AppenderConsole$new(), propagate = FALSE)
lg$appenders[[1]]$set_layout(LayoutGlue$new())
lg$fatal("test")


# All fields of the LogEvent are available, even custom ones
lg$appenders[[1]]$layout$set_fmt("{logger$name} {level_name}({level}) {caller}: {toupper(msg)} {{custom: {custom}}}")
lg$fatal("test", custom = "foobar")


}
\seealso{
Other Layouts: \code{\link{LayoutDbi}},
  \code{\link{LayoutFormat}}, \code{\link{LayoutJson}},
  \code{\link{LayoutTable}}, \code{\link{Layout}}
}
\concept{Layouts}
