% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{LayoutTable}
\alias{LayoutTable}
\title{Abstract Class for Formatting Data as Tabular Structures}
\description{
Abstract classes are exported for developers that want to extend them, they
are not useful to casual users. \link{LayoutDbi} and \link{LayoutJson} are derived
from LayoutTabel.
}
\section{Fields}{


\describe{
\item{\code{event_values}}{Names of the fields of the \link{LogEvent}
to include in the output object.
}
}
}

\section{Methods}{


\describe{
\item{\code{format_event(event)}}{format a \link{LogEvent}}
}
}

\seealso{
Other Layouts: \code{\link{LayoutDbi}},
  \code{\link{LayoutFormat}}, \code{\link{LayoutGlue}},
  \code{\link{LayoutJson}}, \code{\link{Layout}}
}
\concept{Layouts}
