% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Logger.R
\name{print.ancestry}
\alias{print.ancestry}
\alias{format.ancestry}
\title{Print Ancestry of a Logger}
\usage{
\method{print}{ancestry}(x, color = requireNamespace("crayon", quietly =
  TRUE), ...)

\method{format}{ancestry}(x, color = requireNamespace("crayon", quietly =
  TRUE), ...)
}
\arguments{
\item{x}{any \R Object}

\item{color}{\code{TRUE} or \code{FALSE}: Output with color? Requires the Package
\strong{crayon}}

\item{...}{ignored}
}
\value{
\code{x} (invisibly)
}
\description{
Print the LogEvent propagation structure of a chain of Loggers. Usually you
won't have many chained loggers, but if you do this might proof useful.
}
\examples{
 l1 <- Logger$new("AegonV")
 l2 <- Logger$new("Aerys", parent = l1)
 l3 <- Logger$new("Rheagar", parent = l2, propagate = FALSE)
 l4 <- Logger$new("Aegon", parent = l3)

 print(l4$ancestry)
 unclass(l4$ancestry)

}
