% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_tree.R
\name{print.logger_tree}
\alias{print.logger_tree}
\alias{format.logger_tree}
\title{Print Logger Trees}
\usage{
\method{print}{logger_tree}(x, color = requireNamespace("crayon", quietly = TRUE), ...)

\method{format}{logger_tree}(x, color = FALSE, ...)
}
\arguments{
\item{x}{a \link[=logger_tree]{logger_tree}}

\item{color}{\code{logical} scalar. If \code{TRUE} terminal output is colorized via
the package \pkg{crayon}?}

\item{...}{passed on to \code{\link[cli:tree]{cli::tree()}}}
}
\value{
\code{x} (invisibly)
}
\description{
Print Logger Trees
}
