% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_logging.R
\name{simple_logging}
\alias{simple_logging}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\alias{log_exception}
\alias{threshold}
\alias{console_threshold}
\alias{add_appender}
\alias{remove_appender}
\alias{show_log}
\title{Simple Logging}
\usage{
FATAL(msg, ...)

ERROR(msg, ...)

WARN(msg, ...)

INFO(msg, ...)

DEBUG(msg, ...)

TRACE(msg, ...)

log_exception(...)

threshold(level, target = yog::yog)

console_threshold(level, target = yog::yog$appenders$console)

add_appender(appender, name = NULL, target = yog::yog)

remove_appender(pos, target = yog::yog)

show_log(n = 20, threshold = NA, target = yog::yog)
}
\arguments{
\item{msg, ...}{passed on to \code{\link[base:sprintf]{base::sprintf()}}}

\item{level, threshold}{an \code{integer} or \code{character} scalar, see
\code{getOption("yog.log_levels")} for possible values. For \code{threshold} \code{0}
(\code{"off"}) and \code{NA} (\code{"all"}) are also valid.}

\item{target}{a \link{Logger} or \link{Appender}. Defaults to the root logger.}

\item{appender}{an \code{Appender}}

\item{name}{\code{character} scalar. An optional name for the new Appender.}

\item{pos}{\code{integer} index or \code{character} names of the appenders to remove}

\item{n}{\code{integer} scalar. Show only the last \code{n} log entries that match
\code{threshold}}
}
\value{
\code{FATAL()} ... \code{TRACE()} and \code{log_exception()} return the log message
as a \code{character} vector.

\code{threshold()} and \code{console_threshold()} return the \link{log_level} of \code{target}
as \code{integer} (invisibly)

\code{add_appender()} and \code{remove_appender()} return \code{target}.
}
\description{
These functions provide a simple interface to the root logger. If you do not
need any of the more advanced features of yog, start here.

Yog provides convenience functions to manage the root Logger. These
are intended for interactive use, and for people who just need basic
logging facilities and don't want to worry about hierarchical loggers and
R6 classes.

\code{threshold()} sets or retrieves the threshold for an \link{Appender} or \link{Logger}
(the minimum level of log messages it processes). It's \code{target} defaults
to the root logger.

\code{console_threshold()} is a shortcut to set the threshold of the root
loggers \link{AppenderConsole}, which is usually the only Appender that manages
console output for a given \R session.

\code{add_appender()} and \code{remove_appender()} add Appenders to Loggers and
other Appenders.
}
\examples{
FATAL("This is an important message about \%s going wrong", "something")
DEBUG("Debug messages are hidden by default")
console_threshold("debug")  # you must use lower case names here
DEBUG("Unless we lower the threshold")

add_appender(AppenderConsole$new(), "second_console_appender")
FATAL("Multiple console appenders are a bad idea")
remove_appender("second_console_appender")
INFO("Good that we defined an appender name, so it's easy to remove")
}
