% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_caller}
\alias{get_caller}
\alias{system_infos}
\alias{get_user}
\title{Information About the System}
\usage{
get_caller(where = -1L)

get_user(fallback = "unknown user")
}
\arguments{
\item{where}{\code{integer} scalar (usually negative). Look up that many frames
up the call stack}

\item{fallback}{A fallback in case the user name could not be determined}
}
\value{
a \code{character} scalar.
}
\description{
\code{get_caller()} Tries to determine the calling functions based on \code{where}.
}
\examples{
foo <- function() get_caller(-1L)
foo()
get_user()

}
\seealso{
\code{\link[base:sys.call]{base::sys.call()}}

\code{\link[whoami:whoami]{whoami::whoami()}}
}
